==> 01_gtfs_data_ingestion.Rmd <==
---
title: "Chicago GTFS data ingestion"
author: "jan.zymla"
output: html_notebook
---

```{r}
library(tidyverse)
library(data.table)
library(magrittr)
library(lubridate)
library(purrr)
```
GTFS data feed is available at `https://www.transitchicago.com/downloads/sch_data`
```{r}
library(uuid)
library(reticulate)
library(docker)
library(RPostgres)
```
```{r}
getwd()
```

```{r}
client <- docker$from_env()
```

```{r}
data_folder_path = normalizePath('data_raw', winslash = '\\')
vols <- reticulate::dict()
vols[data_folder_path] <- reticulate::dict(bind = '/extdata/', mode = 'ro')
vols
```
```{r}
ports <- reticulate::dict('5432/tcp' = reticulate::tuple('127.0.0.1', as.integer(1100)))
ports
```
<<<<<<< HEAD
You need to have the `starefossen/pgrouting:latest` docker image available. If you haven't done so already, you need to run in your shell `docker pull starefossen/pgrouting:latest`  beforehand.
=======
```{r}
pwd <- uuid::UUIDgenerate() %>% as.character()
env <- reticulate::dict('POSTGRES_PASSWORD' = pwd)
env
```

## Start the container
>>>>>>> 7653b471b4edcf3d472554c9bcdc308e34c7de1c
```{r}
pg <- client$containers$run(name = 'pg', image = "starefossen/pgrouting:latest", remove = TRUE, ports = ports, volumes = vols, detach = TRUE, environment = env)
```

# Download and extract GTFS data
```{r eval=FALSE}
temp <- tempfile()
download.file('https://www.transitchicago.com/downloads/sch_data/google_transit.zip', temp)
```
```{r eval=FALSE}
unzip(zipfile = temp, exdir = data_folder_path)
```



# Now let's connect to the database
```{r}
pg_host <- 'localhost';
pg_port <- '1100';
pg_database <- 'postgres';
pg_user <- 'postgres';
pg_password <- pwd
```


```{r}
con <- 
  dbConnect(
    Postgres(),
    dbname = pg_database,
    host = pg_host,
    port = pg_port,
    user = pg_user,
    password = pg_password
    )
```


```{sql, connection=con}
DROP TABLE IF EXISTS agency;
```


```{sql, connection=con}
CREATE TABLE agency
(
--  agency_id VARCHAR(100) NOT NULL PRIMARY KEY,
  agency_name VARCHAR(100),
  agency_url TEXT,
  agency_timezone VARCHAR(100),
  agency_lang VARCHAR(50),
  agency_phone VARCHAR(50),
  agency_fare_url VARCHAR(100)
)
```

```{sql, connection=con}
COPY agency(agency_name, agency_url, agency_timezone, agency_lang, agency_phone, agency_fare_url) 
FROM '/extdata/agency.txt' DELIMITER ',' CSV HEADER;
```
```{sql connection=con}
CREATE TABLE calendar
(
  service_id VARCHAR(100) NOT NULL,
  monday SMALLINT,
  tuesday SMALLINT,
  wednesday SMALLINT,
  thursday SMALLINT,
  friday SMALLINT,
  saturday SMALLINT,
  sunday SMALLINT,
  start_date DATE,
  end_date DATE,
  start_date_str CHAR(8),
  end_date_str CHAR(8));
```
```{sql, connection=con}
COPY calendar(service_id, monday, tuesday, wednesday, thursday, friday, saturday, sunday, start_date_str, end_date_str) 
FROM '/extdata/calendar.txt' DELIMITER ',' CSV HEADER;
```

```{sql connection=con}
UPDATE calendar SET start_date = to_date(start_date_str, 'YYYYMMDD')::date;
```
```{sql connection=con}
UPDATE calendar SET end_date = to_date(end_date_str, 'YYYYMMDD')::date;
```
```{sql connection = con}
CREATE TABLE calendar_dates
(
  service_id VARCHAR(100) NOT NULL,
  date DATE,
  exception_type SMALLINT,
  date_str CHAR(8)
  );
```

```{sql connection = con}
COPY calendar_dates(service_id, date_str, exception_type) 
FROM '/extdata/calendar_dates.txt' DELIMITER ',' CSV HEADER;
```

```{sql connection = con}
UPDATE calendar_dates SET date = to_date(date_str, 'YYYYMMDD')::date;
```
```{sql connection = con}
DROP TABLE IF EXISTS stop_times;
```
```{sql connection = con}
CREATE TABLE stop_times
(
  trip_id VARCHAR(100) NOT NULL,
  arrival_time interval,
  departure_time interval,
  stop_id VARCHAR(100),
  stop_sequence INT,
  stop_headsign VARCHAR(50),
  pickup_type SMALLINT,
  shape_dist_traveled INT);
```
```{sql connection = con}
COPY stop_times(trip_id,arrival_time,departure_time,stop_id,stop_sequence,stop_headsign,pickup_type,shape_dist_traveled) 
FROM '/extdata/stop_times.txt' DELIMITER ',' CSV HEADER;
```
```{sql connection = con}
CREATE TABLE routes
(
  route_id VARCHAR(100) NOT NULL PRIMARY KEY,
-- agency_id VARCHAR(100),
  route_short_name VARCHAR(100),
  route_long_name VARCHAR(100),
--  route_desc VARCHAR(100),
  route_type SMALLINT,
  route_url VARCHAR(100),
  route_color VARCHAR(100),
  route_text_color VARCHAR(100));
```
```{sql connection = con}
COPY routes(route_id, route_short_name, route_long_name, route_type, route_url, route_color, route_text_color) 
FROM '/extdata/routes.txt' DELIMITER ',' CSV HEADER;
```
```{sql connection = con}
DROP TABLE IF EXISTS stops;
```
```{sql connection = con}
CREATE TABLE stops
(
  stop_id VARCHAR(100) NOT NULL PRIMARY KEY,
  stop_code VARCHAR(100),
  stop_name VARCHAR(100),
  stop_desc VARCHAR(100),
  stop_lat NUMERIC,
  stop_long NUMERIC,
--  zone_id VARCHAR(100),
--  stop_url VARCHAR(100),
  location_type SMALLINT,
  parent_station VARCHAR(100),
  wheelchair_boarding VARCHAR(100));
```
```{sql connection = con}
COPY stops(stop_id, stop_code, stop_name, stop_desc, stop_lat, stop_long, location_type, parent_station, wheelchair_boarding) 
FROM '/extdata/stops.txt' DELIMITER ',' CSV HEADER;
```
```{sql connection = con}
DROP TABLE IF EXISTS transfers;
```
```{sql connection = con}
CREATE TABLE transfers
(
  from_stop_id VARCHAR(100) NOT NULL,
  to_stop_id VARCHAR(100) NOT NULL,
  transfer_type SMALLINT,
--  min_transfer_time  SMALLINT,
  PRIMARY KEY(from_stop_id, to_stop_id)
);
```
```{sql connection = con}
COPY transfers(from_stop_id, to_stop_id, transfer_type) 
FROM '/extdata/transfers.txt' DELIMITER ',' CSV HEADER;
```
```{sql connection = con}
CREATE FUNCTION seconds_to_interval(int) RETURNS interval
    AS $$ SELECT make_interval(secs => $1) $$
    LANGUAGE SQL;
```
```{sql connection = con}
ALTER TABLE transfers ADD COLUMN min_transfer_time interval;
```
```{sql connection = con}
UPDATE transfers SET min_transfer_time = seconds_to_interval(60);
```
```{sql connection = con}
DROP TABLE IF EXISTS trips;
```
```{sql connection = con}
CREATE TABLE trips(
  route_id VARCHAR(100),
  service_id VARCHAR(100),
  trip_id  VARCHAR(100) NOT NULL PRIMARY KEY,
--  trip_headsign  VARCHAR(100),
--  trip_short_name VARCHAR(100),
  direction_id SMALLINT,
  block_id VARCHAR(100),
  shape_id VARCHAR(100),
  direction VARCHAR(50),
  wheelchair_accessible SMALLINT,
--  bikes_allowed SMALLINT,
--  trip_desc VARCHAR(100),
  schd_trip_id  VARCHAR(100)
  );
```
```{sql connection = con}
COPY trips(route_id,service_id,trip_id,direction_id,block_id,shape_id,direction,wheelchair_accessible,schd_trip_id) 
FROM '/extdata/trips.txt' DELIMITER ',' CSV HEADER;
```

```{python eval=FALSE}
import pandas as pd
route_type_names = pd.DataFrame(data={'route_type': [0, 1, 2, 3, 7], 'route_type_name': ['tram', 'metro', 'rail', 'bus', 'funicular']})
route_type_names.reset_index().drop('index', 1)
route_type_names
```
```{r}
route_type_names <- 
  tribble(
    ~route_type, ~route_type_name,
    0, 'tram',
    1, 'metro',
    2, 'rail',
    3, 'bus',
    7, 'funicular'
  )
route_type_names
```
```{r}
dbWriteTable(
  con, 
  'route_type_names',
  route_type_names, 
  overwrite = TRUE, 
  field.types = c('route_type' = 'SMALLINT', 'route_type_name' = 'VARCHAR(100)')
  )
```


# GTFS data processing
```{sql connection = con}
CREATE TABLE routes_rail AS
SELECT routes.*, route_type_name FROM
routes JOIN route_type_names ON routes.route_type = route_type_names.route_type
WHERE route_type_names.route_type_name != 'bus';
```
## calender_dates
exception_type	Required	The exception_type indicates whether service is available on the date specified in the date field.
* A value of 1 indicates that service has been added for the specified date.
* A value of 2 indicates that service has been removed for the specified date.
For example, suppose a route has one set of trips available on holidays and another set of trips available on all other days. You could have one service_id that corresponds to the regular service schedule and another service_id that corresponds to the holiday schedule. For a particular holiday, you would use the calendar_dates.txt file to add the holiday to the holiday service_id and to remove the holiday from the regular service_id schedule.
```{sql connection = con}
DROP TABLE IF EXISTS trips_rail_tuesday;
```
```{sql connection = con}
CREATE TABLE trips_rail_tuesday AS
SELECT trips_rail.* FROM
(
	SELECT
		trips.*,
		route_short_name,
		route_long_name,
		route_color,
		route_text_color,
        route_type_name
	FROM trips
	INNER JOIN routes_rail
	ON trips.route_id = routes_rail.route_id
 ) trips_rail INNER JOIN 
(SELECT coalesce(calendar.service_id, calendar_dates.service_id) AS service_id -- date(date_trunc('week', date '2018-11-18' + 5)) + 1, calendar.service_id, calendar_dates.service_id, *
FROM calendar FULL JOIN calendar_dates
ON calendar.service_id = calendar_dates.service_id AND date(date_trunc('week', date '2018-11-12' + 5)) + 1 = calendar_dates.date
 WHERE	(date(date_trunc('week', date '2018-11-18' + 5)) + 1 BETWEEN start_date AND end_date + 1
 			AND tuesday = 1
			AND (exception_type IS NULL OR exception_type = 1)
		) 
	OR (exception_type = 1 AND date = date(date_trunc('week', date '2018-11-18' + 5)))) services 
ON trips_rail.service_id = services.service_id;
```
```{sql connection = con}
DROP TABLE IF EXISTS stop_times_rail_tuesday;
```
```{sql connection = con}
CREATE TABLE stop_times_rail_tuesday AS
SELECT 
    stops.stop_id,
    stop_name,
    stop_lat,
    stop_long,
	trips_rail_tuesday.*,
	arrival_time,
	departure_time,
	stop_sequence,
	pickup_type -- ,
--	drop_off_type
FROM trips_rail_tuesday
INNER JOIN stop_times ON trips_rail_tuesday.trip_id = stop_times.trip_id
LEFT JOIN stops ON stops.stop_id = stop_times.stop_id;
```
```{sql connection = con}
DROP TABLE IF EXISTS nodes_gtfs;
```
## Nodes
```{sql connection = con}
CREATE TABLE nodes_gtfs AS
SELECT
	trip_id||'::'||stop_sequence||'::dep' AS node_id,
	departure_time||' ('||route_short_name||') '||stop_name||' ->' AS node_name,
	'stop_departure' AS node_type,
	departure_time AS node_time,
	stop_id,
	stop_name,
	stop_lat AS lat,
	stop_long AS lon,
	route_id,
--	route_short_name,
--	route_long_name,
    route_type_name,
	route_color,
	route_text_color,
    trip_id,
    stop_sequence -- ,
--	trip_short_name,
--	trip_headsign
	FROM stop_times_rail_tuesday
UNION ALL
SELECT
	trip_id||'::'||stop_sequence||'::arr' AS node_id,
	arrival_time||' ('||route_short_name||') '||stop_name||' <-' AS node_name,
	'stop_arrival' AS node_type,
	arrival_time AS node_time,
	stop_id,
	stop_name,
	stop_lat AS lat,
	stop_long AS lon,
	route_id,
--	route_short_name,
--	route_long_name,
    route_type_name,
	route_color,
	route_text_color,
    trip_id,
    stop_sequence --,
--	trip_short_name,
--	trip_headsign
	FROM stop_times_rail_tuesday
UNION ALL
SELECT DISTINCT ON (stop_id)
	stop_id||'::entry' AS node_id,
	'entry_node : ' || stop_name AS node_name,
	'entry_node' AS node_type,
	CAST(NULL AS INTERVAL) AS node_time,
	stop_id,
	stop_name,
	stop_lat AS lat,
	stop_long AS lon,
	CAST(NULL AS TEXT) AS route_id,
--	CAST(NULL AS TEXT) AS route_short_name,
--	CAST(NULL AS TEXT) AS route_long_name,
    'walk' AS route_type_name,
	CAST(NULL AS TEXT) AS route_color,
	CAST(NULL AS TEXT) AS route_text_color,
    CAST(NULL AS TEXT) AS trip_id,
    CAST(NULL AS SMALLINT) AS stop_sequence --,
--	CAST(NULL AS TEXT) AS trip_short_name,
--	CAST(NULL AS TEXT) AS trip_headsign
	FROM (
	  SELECT stop_id, stop_name, stop_lat, stop_long
	  FROM stop_times_rail_tuesday
	  GROUP BY stop_id, stop_name, stop_lat, stop_long 
	  ) stop_times_rail_tuesday_distinct_entry
UNION ALL
SELECT
	stop_id||'::exit' AS node_id,
	'exit_node : ' || stop_name AS node_name,
	'exit_node' AS node_type,
	CAST(NULL AS INTERVAL) AS node_time,
	stop_id,
	stop_name,
	stop_lat AS lat,
	stop_long AS lon,
	CAST(NULL AS TEXT) AS route_id,
--	CAST(NULL AS TEXT) AS route_short_name,
--	CAST(NULL AS TEXT) AS route_long_name,
    'walk' AS route_type_name,
	CAST(NULL AS TEXT) AS route_color,
	CAST(NULL AS TEXT) AS route_text_color,
    CAST(NULL AS TEXT) AS trip_id,
    CAST(NULL AS SMALLINT) AS stop_sequence -- ,
--	CAST(NULL AS TEXT) AS trip_short_name,
--	CAST(NULL AS TEXT) AS trip_headsign
	FROM (
	  SELECT stop_id, stop_name, stop_lat, stop_long
	  FROM stop_times_rail_tuesday
	  GROUP BY stop_id, stop_name, stop_lat, stop_long 
	  ) stop_times_rail_tuesday_distinct_exit
;
```
```{sql connection = con}
ALTER TABLE nodes_gtfs ADD id serial PRIMARY KEY
```

```{sql connection = con}
ALTER TABLE nodes_gtfs ADD  geog geometry
```
```{sql connection = con}
UPDATE nodes_gtfs
SET geog = ST_SetSRID(ST_POINT(lon, lat),4326)
```
```{sql connection = con}

```
```{sql connection = con}
```
```{sql connection = con}
```
```{sql connection = con}
```





```{sql connection = con}
DROP TABLE IF EXISTS edges_gtfs;
```
## Edges
```{sql connection = con}
CREATE TABLE edges_gtfs AS
SELECT
  nodes_stop_arrivals.id AS source,
  nodes_stop_departures.id AS target,
	'wait_aboard' AS edge_type,
	nodes_stop_arrivals.node_id AS from_node_id,
	nodes_stop_arrivals.node_name AS from_node_name,
	nodes_stop_arrivals.node_time AS from_node_time,
	nodes_stop_arrivals.lat AS from_lat,
	nodes_stop_arrivals.lon AS from_lon,
	nodes_stop_departures.node_id AS to_node_id,
	nodes_stop_departures.node_name AS to_node_name,
	nodes_stop_departures.node_time AS to_node_time,
	nodes_stop_departures.lat AS to_lat,
	nodes_stop_departures.lon AS to_lon,
	CAST(NULL AS TEXT) AS route_color,
	EXTRACT(EPOCH FROM (nodes_stop_departures.node_time - nodes_stop_arrivals.node_time)) AS cost,
	99999::INTEGER AS reverse_cost
FROM (
        SELECT * 
        FROM nodes_gtfs
        WHERE node_type = 'stop_departure'
        ) nodes_stop_departures
JOIN (
        SELECT * 
        FROM nodes_gtfs
        WHERE node_type = 'stop_arrival'
        ) nodes_stop_arrivals
ON nodes_stop_departures.trip_id = nodes_stop_arrivals.trip_id
AND nodes_stop_departures.stop_sequence = nodes_stop_arrivals.stop_sequence

UNION ALL

SELECT
  nodes_stop_departures.id AS source,
  nodes_stop_arrivals.id AS target,
	nodes_stop_departures.route_type_name AS edge_type,
	nodes_stop_departures.node_id AS from_node_id,
	nodes_stop_departures.node_name AS from_node_name,
	nodes_stop_departures.node_time AS from_node_time,
	nodes_stop_departures.lat AS from_lat,
	nodes_stop_departures.lon AS from_lon,
	nodes_stop_arrivals.node_id AS to_node_id,
	nodes_stop_arrivals.node_name AS to_node_name,
	nodes_stop_arrivals.node_time AS to_node_time,
	nodes_stop_arrivals.lat AS to_lat,
	nodes_stop_arrivals.lon AS to_lon,
	nodes_stop_departures.route_color AS route_color,
	EXTRACT(EPOCH FROM (nodes_stop_arrivals.node_time - nodes_stop_departures.node_time)) AS cost,
	99999::INTEGER AS reverse_cost
FROM (
        SELECT * 
        FROM nodes_gtfs
        WHERE node_type = 'stop_departure'
        ) nodes_stop_departures
INNER JOIN (
        SELECT * 
        FROM nodes_gtfs
        WHERE node_type = 'stop_arrival'
        ) nodes_stop_arrivals
ON nodes_stop_departures.trip_id = nodes_stop_arrivals.trip_id
AND nodes_stop_departures.stop_sequence + 1 = nodes_stop_arrivals.stop_sequence

UNION ALL

SELECT
  nodes_stop_arrivals.id AS source,
  nodes_stop_departures.id AS target,
	'transfer_same_stop' AS edge_type,
	nodes_stop_arrivals.node_id AS from_node_id,
	nodes_stop_arrivals.node_name AS from_node_name,
	nodes_stop_arrivals.node_time AS from_node_time,
	nodes_stop_arrivals.lat AS from_lat,
	nodes_stop_arrivals.lon AS from_lon,
	nodes_stop_departures.node_id AS to_node_id,
	nodes_stop_departures.node_name AS to_node_name,
	nodes_stop_departures.node_time AS to_node_time,
	nodes_stop_departures.lat AS to_lat,
	nodes_stop_departures.lon AS to_lon,
	CAST(NULL AS TEXT) AS route_color,
	EXTRACT(EPOCH FROM (nodes_stop_departures.node_time - nodes_stop_arrivals.node_time)) AS cost,
	99999::INTEGER AS reverse_cost
FROM (
        SELECT * 
        FROM nodes_gtfs
        WHERE node_type = 'stop_departure'
        ) nodes_stop_departures
INNER JOIN (
        SELECT * 
        FROM nodes_gtfs
        WHERE node_type = 'stop_arrival'
        ) nodes_stop_arrivals
    ON nodes_stop_departures.stop_id = nodes_stop_arrivals.stop_id
	AND nodes_stop_departures.node_time >= nodes_stop_arrivals.node_time + '30 secs'
	AND nodes_stop_departures.node_time <= nodes_stop_arrivals.node_time + '1 hour'
    
UNION ALL
SELECT
  nodes_stop_arrivals.id AS source,
  nodes_stop_departures.id AS target,
	'transfer' AS edge_type,
	nodes_stop_arrivals.node_id AS from_node_id,
	nodes_stop_arrivals.node_name AS from_node_name,
	nodes_stop_arrivals.node_time AS from_node_time,
	nodes_stop_arrivals.lat AS from_lat,
	nodes_stop_arrivals.lon AS from_lon,
	nodes_stop_departures.node_id AS to_node_id,
	nodes_stop_departures.node_name AS to_node_name,
	nodes_stop_departures.node_time AS to_node_time,
	nodes_stop_departures.lat AS to_lat,
	nodes_stop_departures.lon AS to_lon,
	CAST(NULL AS TEXT) AS route_color,
	EXTRACT(EPOCH FROM (nodes_stop_departures.node_time - nodes_stop_arrivals.node_time)) AS cost,
	99999::INTEGER AS reverse_cost
FROM (
        SELECT * 
        FROM nodes_gtfs
        WHERE node_type = 'stop_arrival'
        ) nodes_stop_arrivals 
    JOIN transfers ON nodes_stop_arrivals.stop_id = transfers.from_stop_id
    JOIN (
        SELECT * 
        FROM nodes_gtfs
        WHERE node_type = 'stop_departure'
        ) nodes_stop_departures
        ON nodes_stop_departures.stop_id = transfers.to_stop_id
WHERE nodes_stop_departures.node_time >= nodes_stop_arrivals.node_time + min_transfer_time
	AND nodes_stop_departures.node_time <= nodes_stop_arrivals.node_time + '1 hour'

UNION ALL

SELECT
  nodes_entry.id AS source,
  nodes_stop_departures.id AS target,
	'entry_edge' AS edge_type,
	nodes_entry.node_id AS from_node_id,
	nodes_entry.node_name AS from_node_name,
	nodes_entry.node_time AS from_node_time,
	nodes_entry.lat AS from_lat,
	nodes_entry.lon AS from_lon,
	nodes_stop_departures.node_id AS to_node_id,
	nodes_stop_departures.node_name AS to_node_name,
	nodes_stop_departures.node_time AS to_node_time,
	nodes_stop_departures.lat AS to_lat,
	nodes_stop_departures.lon AS to_lon,
	CAST(NULL AS TEXT) AS route_color,
	0::INTEGER AS cost,
	99999::INTEGER AS reverse_cost
FROM (
        SELECT * 
        FROM nodes_gtfs
        WHERE node_type = 'stop_departure'
        ) nodes_stop_departures
JOIN (
        SELECT * 
        FROM nodes_gtfs
        WHERE node_type = 'entry_node'
        ) nodes_entry ON nodes_stop_departures.stop_id = nodes_entry.stop_id
UNION ALL
SELECT
  nodes_stop_arrivals.id AS source,
  nodes_exit.id AS target,
	'exit_edge' AS edge_type,
	nodes_stop_arrivals.node_id AS from_node_id,
	nodes_stop_arrivals.node_name AS from_node_name,
	nodes_stop_arrivals.node_time AS from_node_time,
	nodes_stop_arrivals.lat AS from_lat,
	nodes_stop_arrivals.lon AS from_lon,
	nodes_exit.node_id AS to_node_id,
	nodes_exit.node_name AS to_node_name,
	nodes_exit.node_time AS to_node_time,
	nodes_exit.lat AS to_lat,
	nodes_exit.lon AS to_lon,
	CAST(NULL AS TEXT) AS route_color,
	0::INTEGER AS cost,
	99999::INTEGER AS reverse_cost
FROM (
        SELECT * 
        FROM nodes_gtfs
        WHERE node_type = 'stop_arrival'
        ) nodes_stop_arrivals 
JOIN (
        SELECT * 
        FROM nodes_gtfs
        WHERE node_type = 'exit_node'
        ) nodes_exit ON nodes_stop_arrivals.stop_id = nodes_exit.stop_id


```
```{sql connection = con}
ALTER TABLE edges_gtfs ADD id serial PRIMARY KEY
```
```{sql connection = con}
ALTER TABLE edges_gtfs ADD  geom geometry
```
```{sql connection = con}
UPDATE edges_gtfs
SET geom = ST_SetSRID(ST_MakeLine(ST_POINT(from_lon, from_lat)::geometry, ST_POINT(to_lon, to_lat)::geometry),4326)
```
```{sql connection = con}
SELECT edge_type, count(*)
FROM edges_gtfs
GROUP BY edge_type
LIMIT 10
```

```{sql connection = con}
SELECT *
FROM edges_gtfs
LIMIT 10
```

```{sql connection = con}
SELECT edge_type, count(*)
FROM edges_gtfs
GROUP BY edge_type
```
```{sql connection = con}
SELECT node_type, count(*)
FROM nodes_gtfs
GROUP BY node_type
```

```{sql connection = con}
SELECT *
FROM nodes_gtfs
WHERE node_type = 'entry_node'
LIMIT 10
```

```{sql connection = con}
SELECT *
FROM nodes_gtfs
WHERE node_type = 'exit_node'
ORDER BY stop_name
-- LIMIT 100
```
```{sql connection = con}
SELECT *
FROM nodes_gtfs
WHERE node_type = 'stop_departure' AND stop_name = 'O''Hare' AND node_time > '19:30'
ORDER BY node_time 
LIMIT 10
```
```{sql connection = con}
SELECT *
FROM nodes_gtfs
WHERE node_type = 'exit_node' AND stop_name LIKE 'Clinton%'

```
```{sql connection = con}
SELECT X.*
FROM pgr_Dijkstra(
	'SELECT id, source, target, cost, reverse_cost FROM edges_gtfs',
	29070,
    105552,
	FALSE
	) AS X
ORDER BY seq;
```
```{sql connection = con}
SELECT r.*, edges_gtfs.*
FROM edges_gtfs
INNER JOIN (SELECT X.*
FROM pgr_Dijkstra(
	'SELECT id, source, target, cost, reverse_cost FROM edges_gtfs',
	29070,
    105552,
    TRUE
	) AS X
ORDER BY seq) r ON edges_gtfs.id = r.edge
```
```{sql connection = con}
SELECT *
FROM edges_gtfs
WHERE target = 105552
```
```{sql connection = con}
SELECT *
FROM stop_times_rail_tuesday
WHERE trip_id = '58139913806'
ORDER BY stop_sequence
```
```{sql connection = con}
SELECT * FROM nodes_gtfs WHERE stop_name = 'O''Hare' AND node_time = '19:35:30'
```
```{sql connection = con}
SELECT * FROM nodes_gtfs WHERE stop_name = 'LaSalle' AND node_time = '18:51:00'
```

```{sql connection = con}
SELECT r.*, edges_gtfs.*
FROM edges_gtfs
INNER JOIN (SELECT X.*
FROM pgr_Dijkstra(
	'SELECT id, source, target, cost, reverse_cost FROM edges_gtfs',
	29050,
    81656,
    TRUE
	) AS X
ORDER BY seq) r ON edges_gtfs.id = r.edge
```
```{sql connection = con}
SELECT *
FROM nodes_gtfs
WHERE  stop_name = 'LaSalle' AND node_type = 'entry_node'
```
```{sql connection = con}
SELECT r.*, edges_gtfs.*
FROM edges_gtfs
INNER JOIN (SELECT X.*
FROM pgr_Dijkstra(
	'SELECT id, source, target, cost, reverse_cost FROM edges_gtfs',
	105428,
    81656,
    TRUE
	) AS X
ORDER BY seq) r ON edges_gtfs.id = r.edge
```

```{sql connection = con}
SELECT *
FROM nodes_gtfs
WHERE  stop_name = 'Midway' AND node_type = 'entry_node'
```
```{sql connection = con, output.var = "route"}
SELECT r.*, edges_gtfs.*
FROM edges_gtfs
INNER JOIN (SELECT X.*
FROM pgr_Dijkstra(
	'SELECT id, source, target, cost, reverse_cost FROM edges_gtfs',
	105347,
    81656,
    TRUE
	) AS X
ORDER BY seq) r ON edges_gtfs.id = r.edge
```

```{r}
route
```
#Leaflet
```{r}
library(leaflet)
```
```{r}
leaflet(route %>% mutate(route_color = str_c('#', route_color), lon = map2(from_lon, to_lon, function(x,y) c(x,y)), lat = map2(from_lat, to_lat, function(xx,yy) c(xx,yy)))) %>% 
  setView(lng = -88, lat = 42, zoom = 9) %>% 
  addProviderTiles(providers$Stamen.Toner) %>% 
#  addCircles(lng = ~from_lon, lat = ~from_lat, weight = 1, radius = 1, color = ~route_color) %>% 
  addPolylines(
    data = route %>% 
      mutate(
        route_color = str_c('#', route_color),
#        lon = map2(from_lon, to_lon, function(x,y) c(x,y)),
#        lat = map2(from_lat, to_lat, function(x,y) c(x,y))
        ) %>% 
      select(seq, route_color, from_lon, to_lon, from_lat, to_lat) %>% 
      gather('fto','lon', from_lon:to_lon) %>% 
      gather('fta','lat', from_lat:to_lat) %>% 
      mutate(fta = str_sub(fta, 1, 2), fto = str_sub(fto, 1, 2)) %>% 
      filter(fta == fto) %>% 
      arrange(seq)
  ,
    lng = ~lon, lat = ~lat, group=~seq, fillColor = ~route_color)
```
```{r}
route %>%
  mutate(
    route_color = str_c("#", route_color),
    #        lon = map2(from_lon, to_lon, function(x,y) c(x,y)),
    #        lat = map2(from_lat, to_lat, function(x,y) c(x,y))
  ) %>%
  select(seq, route_color, from_lon, to_lon, from_lat, to_lat) %>%
  gather("fto", "lon", from_lon:to_lo9n) %>%
  gather("fta", "lat", from_lat:to_lat) %>%
  mutate(fta = str_sub(fta, 1, 2), fto = str_sub(fto, 1, 2)) %>%
  filter(fta == fto) %>%
  arrange(seq)
```



```{sql connection = con}
SELECT *
FROM nodes_gtfs
LIMIT 10
```
```{sql connection = con}
```
```{sql connection = con}
```
```{sql connection = con}
```
```{sql connection = con}
```
```{sql connection = con}
```
```{sql connection = con}
```
```{sql connection = con}
```
```{sql connection = con}
```
```{sql connection = con}
```
```{sql connection = con}
```
```{sql connection = con}
```
```{sql connection = con}
```
```{sql connection = con}
```
```{sql connection = con}
```
```{sql connection = con}
```
