==> 10_Clustering stations.Rmd <==
---
title: "10_Clustering stations"
author: "Magda Bolewska"
date: "17 janvier 2019"
output: word_document
---

```{r}
library(tidyverse)
library(lubridate)
library(magrittr)
library(data.table)
library(leaflet)
library(tictoc)
```

## Clustering des stations
Il s'agit d'une première version du clustering des stations (cf. étude 1 du doc "Etat des lieux").

Le clustering est réalisé à partir des données du mardi 04/07/2017.

Etapes à suivre : 
A priori les stations des différents clusters sont assez mélangées :/

1. Tester des différents nombre de clusters en faisant varier la valeur du paramètre km_cent.
Est-ce qu'une de partitions peut s'expliquer par la distribution des infrastructures dans la ville ?

2. Si on change le seed ou ajoute l'option nstart, les résultats sont-ils différents ?

Ensuite : pour l'instant le clustering est fait sur 1 jour ouvré, tester sur un dimanche, essayer d'élargir la période et généraliser le code.

## 1 - Importation des données et traitements préliminaires
```{r}
rm(stations)
tic()
stations_filename <- list.files('data_raw', pattern = 'divvy_stations_[0-9]{4}(_[0-9]{2}){2}_.*\\.csv', full.names = TRUE) %>% max()
stations_filename
stations <- fread(stations_filename, na.strings = c("")) #, nrows = 1E6)
print(toc())
tic()

stations[
  , 
  `:=`(
    available_bikes  = as.integer(available_bikes),
    available_docks  = as.integer(available_docks),
    docks_in_service = as.integer(docks_in_service),
    latitude         = as.double(latitude),
    longitude        = as.double(longitude),
    percent_full     = as.integer(percent_full),
    timestamp        = as_datetime(timestamp),
    total_docks      = as.integer(total_docks)
  ), 
  ]
print(toc())
tic()

stations[,tx_utilisation := available_bikes/docks_in_service]
print(toc())
tic()

stations[,date := as.Date(timestamp)]
print(toc())
tic()

#stations[,heure := format(timestamp, "%T")]
stations[,heure := as.ITime(timestamp)]
print(toc())
tic()

# lubridate::wday is replaced by data.table::wday and a data.table join for higher performance 
#stations[, weekday := lubridate::wday(timestamp, label=TRUE, abbr = FALSE)]
wdays <- 
  (
    seq.Date(ymd('2019-01-21'), ymd('2019-01-28'), 'day') %>% as.data.table()
  )[
    , 
    .(
      wday_n   = data.table::wday(`.`),
      wday_abr = lubridate::wday(`.`, label = TRUE)
      ),
    ][
      , 
      .N, 
      .(wday_n, wday_abr)
    ][
      , 
      .(wday_n, wday_abr)
    ]
stations[, `:=`(wday_n = data.table::wday(timestamp))]
stations <- stations[wdays, on = 'wday_n']
print(toc())
tic()

stations[, month := lubridate::month(timestamp, label=TRUE, abbr = FALSE)]
print(toc())
```
```{r}
seq.Date(ymd('2019-01-21'), ymd('2019-01-28'), 'day') %>% as.data.table() -> a

wdays <- (seq.Date(ymd('2019-01-21'), ymd('2019-01-28'), 'day') %>% as.data.table())[, .(wday_n = data.table::wday(`.`), wday_abr = lubridate::wday(`.`, label = TRUE)),][, .N , .(wday_n, wday_abr)][, .(wday_n, wday_abr)]

```

## 2 - Choix du jour à traiter et analyse de la qualité de données pour cette date
Dans un premier temps, on travaille sur les données du mardi 04/07/2017.
```{r}

# Sélection des stations en service
stations[date == "2017-07-04" & status=="Not In Service",.N,.(id)] # une seule station est concernée, id = 582
#stations2 <- stations[date == "2017-06-06" & id != 582]
stations2 <- stations[wday_n == 3 & id != 582]
length(unique(stations2$id)) # 581 stations en service

# Analyse des "timestamps"
stations2[, .(  'max_period' = max(timestamp - lag(timestamp), na.rm = TRUE),
                'min_period' = min(timestamp - lag(timestamp), na.rm = TRUE)),.(id)][
                  ,.(mi = min(min_period), ma = max(max_period)),]
```
Après vérification : pour l'ensemble de stations nous disposons des relevés réalisés aux mêmes heures, environ toutes les 10 minutes. Un relevé a été réalisé à 20h50 au lieu de 20h55 qui explique le "mi" et "ma".

## 3 - Transformation de la table
Format souhaité : une ligne par station avec des taux d'utilisation à des heures (timestamp) différentes en colonnes
```{r eval=FALSE}
length(unique(stations2$heure)) # 144 

base3 <- stations2[!duplicated(stations2[, id]),.(id)]
liste_stations <- unique(stations2$id)
liste_heures <- unique(stations2$heure)

for(iidd in liste_stations){
  for(h in liste_heures){
    base3[id==iidd,paste("r",gsub(":","_",h),sep="_") := as.double(stations2[id==iidd & heure==h,.(tx_utilisation)])]
  }
}
head(base3)
nrow(base3) # 581
ncol(base3) # 145
```
C'est lent et ça fait honte mais ça marche. Il faut trouver une façon pour faire cette manip sans boucles.

Voilà la version rapide ;)
```{r}
stations3 <- stations2[, .(tx_utilisation = mean(tx_utilisation, na.rm = TRUE)), .(id, hr = hour(timestamp))][hr > 6 | hr < 3]
stations3[, `:=`(hr = as.factor(hr) %>% fct_reorder(hr) %>% fct_shift(3))]
stations3
#base3 <- stations2[, .(id, heure = paste0('r_', str_replace_all(heure, ':', '_')), tx_utilisation)] %>% dcast(id~heure)
base3 <- stations3 %>% dcast(id ~ hr)
base3
```
## 4 - Clustering avec la méthode K-means
```{r}
set.seed(1234)
km_cent <- 3 # PARAMETRE A FAIRE VARIER
classifST <- kmeans(base3[,2:ncol(base3)], centers=km_cent) 
#  rapide, ajouter nstart = 100?

stationsKM <- cbind(base3,classeKM = factor(classifST$cluster))
stationsKM[,.N, by=.(classeKM)]
```

## 5 - Ajout des coordonnées géographiques et visualisation
```{r}
coor_st <- stations[!duplicated(id),.(id,latitude,longitude)]
stations_visu <- merge(stationsKM[,.(id,classeKM)],coor_st,by="id")

factpal <- colorFactor(topo.colors(km_cent), stations_visu$classeKM)
```

```{r}
#stations2[stations_visu[, .(id, classeKM)], on = 'id'] %>% 
stations3[stations_visu[, .(id, classeKM)], on = 'id'] %>% 
  ggplot() +
    geom_point(aes(x = hr, y = tx_utilisation, group = id, color = classeKM), alpha = .1) +
    scale_color_manual(values = factpal(1:km_cent)) +
    facet_wrap(~classeKM) +
    theme(axis.text.x = element_text(angle = 90))
```

```{r}
m <- leaflet::leaflet(data=stations_visu) %>%
  addTiles() %>%
  setView(lng=-87.6297982, lat=41.8781136, zoom=11) %>%
  addCircleMarkers(~longitude, ~latitude, radius=3, color = ~factpal(classeKM))
m
```

```{r}
stations_visu[id != 622] %>% 
  ggplot() +
    geom_point(data = stations_visu[id != 622, .(longitude, latitude)], aes(longitude, latitude), color = 'gray') +
    geom_point(aes(longitude, latitude, color = classeKM)) +
    scale_color_manual(values = factpal(1:km_cent)) +
    facet_wrap(~classeKM)
```

## 6 - Clustering avec la méthode CAH
```{r}
set.seed(1234)
CAH_cent <- 3 # PARAMETRE A FAIRE VARIER
library(cluster)
classifST2 <- agnes(base3[,2:ncol(base3)], method="ward") # method : PARAMETRE A FAIRE VARIER
stationsCAH <- cbind(base3,classeCAH = factor(cutree(classifST2,k=CAH_cent)))
stationsCAH[,.N, by=.(classeCAH)]
```

# Ajout des coordonnées géographiques et visualisation
```{r}
coor_st <- stations[!duplicated(id),.(id,latitude,longitude)]
stations_visu <- merge(stationsCAH[,.(id,classeCAH)],coor_st,by="id")
factpal <- colorFactor(topo.colors(km_cent), stations_visu$classeCAH)
```

```{r} 
stations3[stations_visu[, .(id, classeCAH)], on = 'id'] %>% 
  ggplot() +
    geom_point(aes(x = hr, y = tx_utilisation, group = id, color = classeCAH), alpha = .1) +
    scale_color_manual(values = factpal(1:km_cent)) +
    facet_wrap(~classeCAH) +
    theme(axis.text.x = element_text(angle = 90))
```

```{r}
m <- leaflet::leaflet(data=stations_visu) %>%
  addTiles() %>%
  setView(lng=-87.6297982, lat=41.8781136, zoom=11) %>%
  addCircleMarkers(~longitude, ~latitude, radius=3, color = ~factpal(classeCAH))
m
```

```{r}
stations_visu[id != 622] %>% 
  ggplot() +
    geom_point(data = stations_visu[id != 622, .(longitude, latitude)], aes(longitude, latitude), color = 'gray') +
    geom_point(aes(longitude, latitude, color = classeCAH)) +
    scale_color_manual(values = factpal(1:km_cent)) +
    facet_wrap(~classeCAH)
```
