==> 11_Clustering stations based on trip data.Rmd <==
---
title: "11_Clustering stations based on trip data"
author: "jan.zymla@renault.com"
date: "2019-01-29"
---

```{r}
library(tidyverse)
library(lubridate)
library(magrittr)
library(data.table)
library(leaflet)
library(tictoc)
```


## 1 - Importation des données et traitements préliminaires
```{r}
tic()
trips_filename <- list.files('data_raw', pattern = 'divvy_trips_[0-9]{4}(_[0-9]{2}){2}_.*\\.csv', full.names = TRUE) %>% max()
trips_filename
trips <- 
  fread(
#    nrows = 1e6,
    trips_filename, 
    select = c('from_latitude', 'from_longitude', 'from_station_id', 'start_time', 'stop_time', 'to_latitude', 'to_longitude', 'to_station_id'), 
    na.strings = c("")
    )
trips[
  , 
  `:=`(
    start_time = as_datetime(start_time),
    stop_time = as_datetime(stop_time),
    from_longitude = as.numeric(from_longitude),
    from_latitude = as.numeric(from_latitude),
    from_station_id = as.integer(from_station_id),
    to_latitude = as.numeric(to_latitude),
    to_longitude = as.numeric(to_longitude),
    to_station_id = as.integer(to_station_id)
  ), 
  ]

wdays <- 
    (
        seq.Date(ymd('2019-01-21'), ymd('2019-01-28'), 'day') %>% as.data.table()
    )[
        , 
        .(
            wday_n   = data.table::wday(`.`),
            wday_abr = lubridate::wday(`.`, label = TRUE)
        ),
        ][
            , 
            .N, 
            .(wday_n, wday_abr)
            ][
                order(wday_n), 
                .(wday_n, wday_abr)
                ]
trips[, `:=`(wday_n = data.table::wday(start_time))]
trips <- trips[wdays, on = 'wday_n']
print(toc())
#trips[, weekday := lubridate::wday(start_time, label=TRUE, abbr = FALSE)]
trips[, start_hr := hour(start_time)]
print(toc())
tic()
```


# Selection et mise en forme des données de `trips`
```{r}
station_trips <- 
  (trips[
    from_station_id != to_station_id,
    .(start_time, stop_time, from_station_id, to_station_id)
    ] %>% 
    { 
      rbind(
        .[,.(station_id = from_station_id, timestamp = start_time, type = 'from')],
        .[,.(station_id = to_station_id, timestamp = stop_time, type = 'to')]
      )
    })
station_trips %>% head() %>% as_tibble()
```
```{r}
coor_st <-
  trips[
    !is.na(from_longitude) & !is.na(from_latitude), 
    .(latitude = last(from_latitude), longitude = last(from_longitude)), 
    .(id = from_station_id)
  ]
coor_st %>% as_tibble()
```

```{r}
station_trips[, .(.N, first_date = min(timestamp), yr = year(min(timestamp))), .(id = station_id)][coor_st, on = 'id'] %>% 
  {
    ggplot() +
      geom_point(data = .[,.(longitude, latitude, N)], aes(longitude, latitude, size = log2(N)), colour = 'gray') +
      geom_point(data = ., aes(longitude, latitude, color = log10(N), size = (N))) +
      facet_wrap(~yr)
}
   
```
```{r}
station_movements <-
  station_trips[
    , 
    .(hr = hour(timestamp), wday_n = data.table::wday(timestamp - 4 * 3600), station_id, type)
  ][
    , 
    .N, 
    .(hr, wday_n, station_id, type)
  ][
    coor_st, 
    on = .(station_id = id), 
    nomatch = 0
  ]
station_movements %>% glimpse()
```
```{r, eval = FALSE}
write_rds(x = station_movements, 'data_gold/station_movements.rds')
```



```{r}
#tibble(hr = 0:23) %>% 
 # mutate(hr = as.factor(hr) %>% fct_shift(5)) 
hrtf <-
    union_all(
      tibble(hr = 0:23) %>% 
        mutate(hr = as.factor(hr) %>% fct_shift(5)) %>% 
        mutate(type = 'from', hrtf = fct_relabel(hr, ~paste0('f', .))), #%>% 
#        mutate(hrtf = fct_reorder(hrtf, as.integer(hr))),
      tibble(hr = 0:23) %>% 
        mutate(hr = as.factor(hr) %>% fct_shift(5)) %>%
        mutate(type = 'to',   hrtf = fct_relabel(hr, ~paste0('t', .))) #%>% 
#        mutate(hrtf = fct_reorder(hrtf, as.integer(hr)))
    ) %>%
  mutate(hrtf = fct_reorder(hrtf, as.numeric(hr))) %>% 
  arrange(hrtf) %>% 
  as.data.table()
hrtf
```



```{r}
station_movements_subset <- 
  (
    station_movements[wday_n %in% c(7)][, .(N = sum(N)), .(hr, station_id, type)] %>%
    { 
      rbind(
        .[type == 'from', .(type, hr, N = -N), station_id],
        .[type == 'to',   .(type, hr, N),      station_id]
      )
    }
  )[hrtf, on = c('hr', 'type')][order(station_id, hrtf)][, .(hr, hrtf, N, dailyN = sum(abs(N))), station_id][, .(N = N/dailyN, dailyN, station_id, hrtf)] 
station_movements_subset %>% glimpse()
```
```{r}
station_movements_kmeans_input <-
  station_movements_subset[!is.na(station_id)] %>% dcast(station_id + dailyN ~ hrtf, fill = 0, value.var = 'N')
station_movements_kmeans_input[, `:=`(dailyN = 1000*(dailyN < 500))]

station_movements_kmeans_input
```    
```{r}
classifST <- kmeans(station_movements_kmeans_input[,2:ncol(station_movements_kmeans_input)], centers = 12) 
#classifST
```
```{r}
stationsKM <- cbind(station_movements_kmeans_input, classeKM = factor(classifST$cluster))
stationsKM
```


```{r}
coor_st[stationsKM[,.(id = station_id, classeKM)], on = 'id', ] %>% 
  {
    ggplot() +
      geom_point(data = .[,.(longitude, latitude)], aes(longitude, latitude), colour = 'gray') +
      geom_point(data = ., aes(longitude, latitude, color = classeKM)) +
      facet_wrap(~classeKM)
}
```


```{r}
station_movements_subset[hrtf, on = 'hrtf'][stationsKM[,.(station_id, classeKM = classeKM)], on = 'station_id'] %>% 
#(dst$station_movements_kmeans_input %>% melt('station_id', variable.name = 'hrtf', value.name = 'N'))[hrtf, on='hrtf'][stationsKM[,.(station_id, classeKM = classeKM)], on = 'station_id'] %>% 
  ggplot() +
    geom_line(
      aes(hr, N, group = interaction(station_id, type), color = classeKM), alpha = .1
    ) +
    geom_line(
      data = classifST$centers %>% melt(value.name = 'N') %>% mutate(classeKM = as.factor(Var1), hrtf = Var2)%>% left_join(hrtf %>% as.tibble(), by = 'hrtf'),
      aes(hr, N, group = interaction(classeKM, type))
    ) +
    facet_wrap(~classeKM)  +ylim(-.2, .2)
```
```{r}
station_movements_subset[hrtf, on = 'hrtf'][stationsKM[,.(station_id, classeKM = classeKM)], on = 'station_id'] %>% 
#(dst$station_movements_kmeans_input %>% melt('station_id', variable.name = 'hrtf', value.name = 'N'))[hrtf, on='hrtf'][stationsKM[,.(station_id, classeKM = classeKM)], on = 'station_id'] %>% 
  ggplot() +
    geom_line(
      aes(hr, N, group = interaction(station_id, type), color = classeKM), alpha = .1
    ) +
    geom_line(
      data = classifST$centers %>% melt(value.name = 'N') %>% mutate(classeKM = as.factor(Var1), hrtf = Var2)%>% left_join(hrtf %>% as.tibble(), by = 'hrtf'),
      aes(hr, N, group = interaction(classeKM, type))
    ) +
    facet_wrap(~classeKM)  +ylim(-.2, .2)

```
