==> 13_Customers_trips_construction_BDD_travail_final.Rmd <==
---
title: "Construction de la base de travail des trips pour l'année 2017"
output:
  html_document:
    df_print: paged
date: "08.02.2019"
---


```{r}
#install.packages("chron")

library(data.table)
library(tictoc)
library(lubridate)
library(geosphere)
library(reshape2)
```

Changer les chemins pour : 
- la lecture des deux tables : tmpfs et trips
- l'écriture, la lecture puis la suppression des 3 tables : stations_15min, stations_4h, stations_h
Change (obligée sinon Rstudio sature en mémoire)
- l'écritude sur le disque des deux tables de travail pour Shiny: trajets et des_trajets

#########################################################################
### Mise en forme des données STATIONS
### Ce qu'on souhaite c'est récupérer les taux d'utilisation des stations
### leurs capacités
#########################################################################


```{r}
tmpfs <- fread("D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/data_raw/divvy_stations_2019_02_22_00_56_01_7741b595ace.csv")
```
Changement de format de deux variables

```{r}
tmpfs[,timestamp := as_datetime(timestamp)] 
tmpfs[,total_docks:=as.integer(total_docks)]
```


Création d'une table annuaire des stations uniformisées 
On contrôle des cas où le nom change dans le temps de même que la localistaion géographique
On conserve pour chaque station les dernières valeurs observées pour la station

```{r}
setorder (tmpfs,id,timestamp)

lstations <-tmpfs[,c("id","latitude","longitude","station_name","timestamp")]
setorderv(lstations,c("id","timestamp"))

setkey(lstations, id)
lstations <- lstations[J(unique(id)), mult = "last"]
lstations[,c("timestamp"):=NULL]
lstations<-unique(lstations)
```


#### Calcul de la distance entre chaque station
#### L'idée est de calculer pour chaque station le nb de stations à 300m, 1km...

```{r}
dist_st <- as.data.frame(lstations[,c("id","latitude","longitude")])

    # distance de Haversine
matH<-distm(dist_st[3:2],dist_st[3:2],fun=distHaversine)
# replace the diagonal with NA
diag(matH) <- NA
rownames(matH) <- dist_st[,c("id")]
colnames(matH) <- dist_st[,c("id")]

rm(dist_st)

# Transposition de la matrice en table
tab_dist <- melt(matH) ##Assuming your data is a matrix. i.e. the people's names are the row and col names.
rm(matH)
tab_dist<- as.data.table(tab_dist)
colnames(tab_dist) <- c("station","station2","distanceHav")
tab_dist<-tab_dist[!is.na(distanceHav)]

tab_dist[,moins2km:=ifelse(distanceHav<2001,1L,0L)]
tab_dist[,moins1km:=ifelse(distanceHav<1001,1L,0L)]
tab_dist[,moins300m:=ifelse(distanceHav<301,1L,0L)]
tab_dist[,nbst_300m:=sum(moins300m),by=c("station")]
tab_dist[,nbst_1km:=sum(moins1km),by=c("station")]
tab_dist[,nbst_2km:=sum(moins2km),by=c("station")]
tab_dist<-tab_dist[,c("station","nbst_300m","nbst_1km","nbst_2km")]
tab_dist<-unique(tab_dist)
```


#### Suppression de variables à ce stade non utiles

```{r}
tmpfs[,c("address","latitude","longitude","location","station_name","status"):=NULL]
```


#### Gestion de la date
### On crée les variables hr, jour, mois qui permettront d'apparier les infos sur les stations
### au fichier des trajets par tranche horaire

```{r}
tmpfs[,hr:=hour(timestamp)]
tmpfs[,jour:=day(timestamp)]
tmpfs[,mois:=month(timestamp)]
```


#### Gestion des heures
###  On va construire des tranches horaires pour apparier l'activité des stations avec les données de trajets
### On prend le plus fin pour être au plus d'une distribution de valeurs continues (15 cela marche bien, car plus fin cela crée de la valeur manquante lors de l'appariemment avec trips)

```{r}
  # Tranches horaires de 15 min
for (i in seq(0,59,by=15)){
  tmpfs[(minute(timestamp)>=0+i & minute(timestamp)<15+i), tr_hor_15:= paste("[",i,"-",i+15,"[")]
}
rm(i)

  # Tranches horaires de 4 heures
for (i in seq(0,23,by=4)){
  tmpfs[(hour(timestamp)>=0+i & hour(timestamp)<4+i), hr4:= paste("[",i,"-",i+4,"[")]
}
rm(i)
```


#### Taux d'utilisation des bornes

```{r}
# total_docks : capacité totale de la station (nb total de places disponibles pour accueillir un vélo). 
# docks_in_service: places en fonctionnement pour accueillir un vélo (= total_docks - pannes)

# available_docks: places libres parmi les places en fonctionnement
# available_bikes: nb de vélos disponibles à la station

# pct_full    = taux de remplissage de la station (available_bikes/total_docks)
# pct_full_av = taux de remplissage de la station si on ne tient compte que des places disponibles (available_bikes/docks_in_service)

tmpfs[,pct_full   :=round((available_bikes/total_docks),2)*100]
tmpfs[,pct_full_av:=round((available_bikes/docks_in_service),2)*100]

tmpfs[,c("timestamp","record","percent_full"):=NULL]

```

```{r}
tmpfs<-unique(tmpfs)
```


# Aggrégation par station, date (mois, jour), et tranches horaires

```{r}
  # Pour la capacité on prend la moyenne : on est sur des tranches horaires de 15 min
stations_15min <- copy(tmpfs)
tmpfs[,c("tr_hor_15"):=NULL]

stations_15min <- stations_15min[,m_total_docks:=mean(total_docks),by=.(id,mois,jour,hr,tr_hor_15)]
stations_15min <- stations_15min[,m_docks_in_service:=mean(docks_in_service),by=.(id,mois,jour,hr,tr_hor_15)]
stations_15min <- stations_15min[,m_available_docks:=mean(available_docks),by=.(id,mois,jour,hr,tr_hor_15)]
stations_15min <- stations_15min[,m_available_bikes:=mean(available_bikes),by=.(id,mois,jour,hr,tr_hor_15)]
stations_15min <- stations_15min[,m_pct_full:=mean(pct_full),by=.(id,mois,jour,hr,tr_hor_15)]
stations_15min <- stations_15min[,m_pct_full_av:=mean(pct_full_av),by=.(id,mois,jour,hr,tr_hor_15)]

stations_15min[,c("hr4","total_docks","docks_in_service","available_docks","available_bikes","pct_full","pct_full_av"):=NULL]

 # On supprime les doublons   
stations_15min <- unique(stations_15min)

saveRDS(stations_15min,"D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/fichier_anissa/stations_15min.rda",compress=FALSE)

rm(stations_15min)

```


```{r}
stations_4h <- copy(tmpfs)
tmpfs[,c("hr4","hr"):=NULL]

stations_4h <- stations_4h[,m_total_docks_4h:=mean(total_docks),by=.(id,mois,jour,hr4)]
stations_4h <- stations_4h[,m_docks_in_service_4h:=mean(docks_in_service),by=.(id,mois,jour,hr4)]
stations_4h <- stations_4h[,m_available_docks_4h:=mean(available_docks),by=.(id,mois,jour,hr4)]
stations_4h <- stations_4h[,m_available_bikes_4h:=mean(available_bikes),by=.(id,mois,jour,hr4)]
stations_4h <- stations_4h[,m_pct_full_4h:=mean(pct_full),by=.(id,mois,jour,hr4)]
stations_4h <- stations_4h[,m_pct_full_av_4h:=mean(pct_full_av),by=.(id,mois,jour,hr4)]

stations_4h[,c("hr","total_docks","docks_in_service","available_docks","available_bikes","pct_full","pct_full_av"):=NULL]
stations_4h <- unique(stations_4h)

saveRDS(stations_4h,"D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/fichier_anissa/stations_4h.rda",compress=FALSE)

rm(stations_4h)

```


```{r}
stations_h <- copy(tmpfs)

stations_h <- stations_h[,m_total_docks_h:=mean(total_docks),by=.(id,mois,jour)]
stations_h <- stations_h[,m_docks_in_service_h:=mean(docks_in_service),by=.(id,mois,jour)]
stations_h <- stations_h[,m_available_docks_h:=mean(available_docks),by=.(id,mois,jour)]
stations_h <- stations_h[,m_available_bikes_h:=mean(available_bikes),by=.(id,mois,jour)]
stations_h <- stations_h[,m_pct_full_h:=mean(pct_full),by=.(id,mois,jour)]
stations_h <- stations_h[,m_pct_full_av_h:=mean(pct_full_av),by=.(id,mois,jour)]

stations_h[,c("total_docks","docks_in_service","available_docks","available_bikes","pct_full","pct_full_av"):=NULL]
stations_h <- unique(stations_h)

saveRDS(stations_h,"D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/fichier_anissa/stations_h.rda",compress=FALSE)

rm(stations_h)
rm(tmpfs)
```


#### Importation du fichier des TRIPS
####
########################################


```{r}
trips <- fread("D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/data_raw/divvy_trips_2019_02_22_00_51_40_774496853fd.csv")
```

```{r}
  # On passe en format date
trips[,`:=`(start_time = as_datetime(start_time),stop_time = as_datetime(stop_time))] 
```

### Filtre sur l'année des trips qui nous intéressent : ceux de 2017

```{r}
trajets <- trips[year(start_time)==2017,]

rm(trips)
```


Suppression des caratéristiques des subscribers 

```{r}
trajets[,c("birth_year","gender"):=NULL]
```


### Création de quelques variables d'intérêt

```{r}
# Part des abonnes dans le trajet

trajets[,abonne:=ifelse(user_type=="Customer",1L,0L)]
trajets[,c("user_type"):=NULL]
```


### Calcul distance entre deux points

```{r}
trajets[,distHav:= distHaversine(trajets[,c("from_longitude","from_latitude")],trajets[,c("to_longitude","to_latitude")])]

trajets[,c("from_latitude","from_longitude","from_location","to_latitude","to_longitude","to_location"):=NULL]
```



#### Gestion de la date
### On considère qu'une journée commence à 5h du matin  
### On va décaler les dates et construire une variable qui nous dira si on est en semaine ou en week-end

```{r}
night_thd = 5

  # Départ: Numéro du jour de la semaine : Levels: (1) Sun < Mon < Tue < Wed < Thu < Fri < Sat (7)
trajets[, wday_n := data.table::wday(start_time - night_thd * 3600)]

  # Départ: Heure
trajets[, start_hr := hour(start_time)]
  # Départ: Jour
trajets[, start_jour := day(start_time)]
  # Départ: Mois
trajets[, start_mois := month(start_time)]


  # Arrivée: Heure
trajets[, stop_hr := hour(stop_time)]
  # Arrivée: Jour
trajets[, stop_jour := day(stop_time)]
  # Arrivée: Mois
trajets[, stop_mois := month(stop_time)]

rm(night_thd)
```


### Nb de trips partants et arrivant de/à chaque station

```{r}
st_trips <- trajets[,c("from_station_id","to_station_id","trip_id","start_jour","start_mois","wday_n","start_time")]

  # En pratique il faudrait recoder le jour dans le mois pour prendre en compte
  # qu'une journée commence à 5 heures. 
  # En pratique c'est 1.5% des obs... mais comme ensuite on agrège sur e jour de la semaine 
  # on peut considérer que c'est mineur


# Nombre de trajets partants et arrivants par jour de la semaine en moyenne
# (on fait la somme sur chaque jour du mois et puis la moyenne par jour de la semaine)
st_trips <- st_trips[,v_from:=.N,by=c("from_station_id","start_mois","start_jour")]
st_trips <- st_trips[,v_to:=.N,by=c("to_station_id","start_mois","start_jour")]

st_trips <- st_trips[,moy_from_trips:=mean(v_from),by=c("from_station_id","wday_n")]
st_trips <- st_trips[,moy_to_trips:=mean(v_to),by=c("to_station_id","wday_n")]

st_trips[,c("trip_id","start_time","start_jour","start_mois","v_from","v_to"):=NULL]
st_trips <- unique(st_trips)
  # 390 231 obs

trajets[,c("trip_id"):=NULL]

```



### Nb de vélos différents partants et arrivant de/à chaque station

```{r}
st_bikes <- trajets[,c("from_station_id","to_station_id","bike_id","start_jour","start_mois","wday_n")]
st_bikes <- unique(st_bikes)

  # En pratique il faudrait recoder le jour dans le mois pour prendre en compte
  # qu'une journée commence à 5 heures. 
  # En pratique c'est 1.5% des obs... mais comme ensuite on agrège sur e jour de la semaine 
  # on peut considérer que c'est mineur


# Nombre de trajets partants et arrivants par jour de la semaine en moyenne
# (on fait la somme sur chaque jour du mois et puis la moyenne par jour de la semaine)
st_bikes <- st_bikes[,v_from:=.N,by=c("from_station_id","start_mois","start_jour")]
st_bikes <- st_bikes[,v_to:=.N,by=c("to_station_id","start_mois","start_jour")]

st_bikes <- st_bikes[,moy_from_bikes:=mean(v_from),by=c("from_station_id","wday_n")]
st_bikes <- st_bikes[,moy_to_bikes:=mean(v_to),by=c("to_station_id","wday_n")]

st_bikes[,c("bike_id","start_jour","start_mois","v_from","v_to"):=NULL]
st_bikes <- unique(st_bikes)
  # 390 231 obs

trajets[,c("bike_id"):=NULL]

```


#### On ajoute les infos sur le nb de station autours

```{r}
trajets <- merge(trajets,tab_dist,by.x=c("from_station_id"),by.y=c("station"),ALL.x=TRUE)

setnames(trajets,"nbst_300m","from_nbst_300m")
setnames(trajets,"nbst_1km","from_nbst_1km")
setnames(trajets,"nbst_2km","from_nbst_2km")

trajets<-merge(trajets,tab_dist,by.x=c("to_station_id"),by.y=c("station"),ALL.x=TRUE)
setnames(trajets,"nbst_300m","to_nbst_300m")
setnames(trajets,"nbst_1km","to_nbst_1km")
setnames(trajets,"nbst_2km","to_nbst_2km")

rm(tab_dist)
```


#### On ajoute : les flux de cyclistes arrivants et partants et le nb de vélos différents sur la journée

```{r}

trajets <- merge(trajets,st_trips,by.x=c("from_station_id","to_station_id","wday_n"),by.y=c("from_station_id","to_station_id","wday_n"),ALL.x=TRUE)

trajets <- merge(trajets,st_bikes,by.x=c("from_station_id","to_station_id","wday_n"),by.y=c("from_station_id","to_station_id","wday_n"),ALL.x=TRUE)

rm(st_bikes)
rm(st_trips)
```



#### Gestion des heures
###  On va construire des tranches de 15 min par heure,
###  notamment pour apparier avec les données stations

```{r}
  # Tranches d'heures de 15 min
for (i in seq(0,59,by=15)){
  trajets[(minute(start_time)>=0+i & minute(start_time)<15+i), start_tr_hor_15:= paste("[",i,"-",i+15,"[")]
  trajets[(minute(stop_time)>=0+i & minute(stop_time)<15+i), stop_tr_hor_15:= paste("[",i,"-",i+15,"[")]
}

  # Tranches de 4 heures
for (i in seq(0,23,by=4)){
  trajets[(hour(start_time)>=0+i & hour(start_time)<4+i), start_hr4:= paste("[",i,"-",i+4,"[")]
  trajets[(hour(stop_time)>=0+i & hour(stop_time)<4+i), stop_hr4:= paste("[",i,"-",i+4,"[")]
}

trajets[,c("start_time","stop_time"):=NULL]
rm(i)
```


```{r}
trajets<-unique(trajets)
```



############################################################
#### On ajoute les infos de stations à la table des trajets
############################################################


```{r}
# On fait un left outer join pour ajouter les variables en fonction des valeurs de départ des trajets

stations_15min <- readRDS("D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/fichier_anissa/stations_15min.rda")

trajets[,id:=from_station_id]
trajets[,mois:=start_mois]
trajets[,jour:=start_jour]
trajets[,hr:=start_hr]
trajets[,tr_hor_15:=start_tr_hor_15]

trajets<-merge(trajets,stations_15min,by=c("id","mois","jour","hr","tr_hor_15"),all.x =TRUE)

setnames(trajets, "m_total_docks","start_m_total_docks")
setnames(trajets, "m_docks_in_service","start_m_docks_in_service")
setnames(trajets, "m_available_docks","start_m_available_docks")
setnames(trajets, "m_available_bikes","start_m_available_bikes")
setnames(trajets, "m_pct_full","start_m_pct_full")
setnames(trajets, "m_pct_full_av","start_m_pct_full_av")

# On fait un left outer join pour ajouter les variables en fonction des valeurs d'arrivée des trajets
trajets[,c("id","mois","jour","hr","tr_hor_15"):=NULL]

trajets[,id:=to_station_id]
trajets[,mois:=stop_mois]
trajets[,jour:=stop_jour]
trajets[,hr:=stop_hr]
trajets[,tr_hor_15:=stop_tr_hor_15]

trajets<-merge(trajets,stations_15min,by=c("id","mois","jour","hr","tr_hor_15"),all.x =TRUE)

setnames(trajets, "m_total_docks","stop_m_total_docks")
setnames(trajets, "m_docks_in_service","stop_m_docks_in_service")
setnames(trajets, "m_available_docks","stop_m_available_docks")
setnames(trajets, "m_available_bikes","stop_m_available_bikes")
setnames(trajets, "m_pct_full","stop_m_pct_full")
setnames(trajets, "m_pct_full_av","stop_m_pct_full_av")

trajets[,c("start_tr_hor_15","stop_tr_hor_15"):=NULL]

rm(stations_15min)
file.remove("D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/fichier_anissa/stations_15min.rda")
```


```{r}
  # on remplace les potentielles NA : d'abord par tranche de 4 heures

stations_4h <- readRDS("D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/fichier_anissa/stations_4h.rda")

trajets[,c("id","mois","jour","hr","tr_hor_15"):=NULL]

trajets[,id:=from_station_id]
trajets[,mois:=start_mois]
trajets[,jour:=start_jour]
trajets[,hr4:=start_hr4]

trajets<-merge(trajets,stations_4h,by=c("id","mois","jour","hr4"),all.x =TRUE)

trajets[,start_m_total_docks:=(ifelse(is.na(start_m_total_docks),m_total_docks_4h,start_m_total_docks))]  
trajets[,start_m_docks_in_service:=(ifelse(is.na(start_m_docks_in_service),m_docks_in_service_4h,start_m_docks_in_service))]  
trajets[,start_m_available_docks:=(ifelse(is.na(start_m_available_docks),m_available_docks_4h,start_m_available_docks))]  
trajets[,start_m_available_bikes:=(ifelse(is.na(start_m_available_bikes),m_available_bikes_4h,start_m_available_bikes))]  
trajets[,start_m_pct_full:=(ifelse(is.na(start_m_pct_full),m_pct_full_4h,start_m_pct_full))]  
trajets[,start_m_pct_full_av:=(ifelse(is.na(start_m_pct_full_av),m_pct_full_av_4h,start_m_pct_full_av))]  

trajets[,c("id","mois","jour","hr4"):=NULL]

trajets[,c("m_total_docks_4h","m_docks_in_service_4h","m_available_docks_4h","m_available_bikes_4h","m_pct_full_4h","m_pct_full_av_4h"):=NULL]

trajets[,id:=to_station_id]
trajets[,mois:=stop_mois]
trajets[,jour:=stop_jour]
trajets[,hr4:=stop_hr4]

trajets<-merge(trajets,stations_4h,by=c("id","mois","jour","hr4"),all.x =TRUE)

trajets[,stop_m_total_docks:=(ifelse(is.na(stop_m_total_docks),m_total_docks_4h,stop_m_total_docks))]  
trajets[,stop_m_docks_in_service:=(ifelse(is.na(stop_m_docks_in_service),m_docks_in_service_4h,stop_m_docks_in_service))]  
trajets[,stop_m_available_docks:=(ifelse(is.na(stop_m_available_docks),m_available_docks_4h,stop_m_available_docks))]  
trajets[,stop_m_available_bikes:=(ifelse(is.na(stop_m_available_bikes),m_available_bikes_4h,stop_m_available_bikes))]  
trajets[,stop_m_pct_full:=(ifelse(is.na(stop_m_pct_full),m_pct_full_4h,stop_m_pct_full))]  
trajets[,stop_m_pct_full_av:=(ifelse(is.na(stop_m_pct_full_av),m_pct_full_av_4h,stop_m_pct_full_av))]  

trajets[,c("m_total_docks_4h","m_docks_in_service_4h","m_available_docks_4h","m_available_bikes_4h","m_pct_full_4h","m_pct_full_av_4h"):=NULL]
trajets[,c("start_hr4","stop_hr4"):=NULL]

rm(stations_4h)
file.remove("D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/fichier_anissa/stations_4h.rda")

```


```{r}
  # on remplace les potentielles NA restantes : valeur moyenne sur la journée

stations_h <- readRDS("D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/fichier_anissa/stations_h.rda")

trajets[,c("id","mois","jour","hr4"):=NULL]

trajets[,id:=from_station_id]
trajets[,mois:=start_mois]
trajets[,jour:=start_jour]

trajets<-merge(trajets,stations_h,by=c("id","mois","jour"),all.x =TRUE)

trajets[,start_m_total_docks:=(ifelse(is.na(start_m_total_docks),m_total_docks_h,start_m_total_docks))]  
trajets[,start_m_docks_in_service:=(ifelse(is.na(start_m_docks_in_service),m_docks_in_service_h,start_m_docks_in_service))]  
trajets[,start_m_available_docks:=(ifelse(is.na(start_m_available_docks),m_available_docks_h,start_m_available_docks))]  
trajets[,start_m_available_bikes:=(ifelse(is.na(start_m_available_bikes),m_available_bikes_h,start_m_available_bikes))]  
trajets[,start_m_pct_full:=(ifelse(is.na(start_m_pct_full),m_pct_full_h,start_m_pct_full))]  
trajets[,start_m_pct_full_av:=(ifelse(is.na(start_m_pct_full_av),m_pct_full_av_h,start_m_pct_full_av))]  

trajets[,c("id","mois","jour"):=NULL]
trajets[,c("m_total_docks_h","m_docks_in_service_h","m_available_docks_h","m_available_bikes_h","m_pct_full_h","m_pct_full_av_h"):=NULL]

trajets[,id:=to_station_id]
trajets[,mois:=stop_mois]
trajets[,jour:=stop_jour]

trajets<-merge(trajets,stations_h,by=c("id","mois","jour"),all.x =TRUE)

trajets[,stop_m_total_docks:=(ifelse(is.na(stop_m_total_docks),m_total_docks_h,stop_m_total_docks))]  
trajets[,stop_m_docks_in_service:=(ifelse(is.na(stop_m_docks_in_service),m_docks_in_service_h,stop_m_docks_in_service))]  
trajets[,stop_m_available_docks:=(ifelse(is.na(stop_m_available_docks),m_available_docks_h,stop_m_available_docks))]  
trajets[,stop_m_available_bikes:=(ifelse(is.na(stop_m_available_bikes),m_available_bikes_h,stop_m_available_bikes))]  
trajets[,stop_m_pct_full:=(ifelse(is.na(stop_m_pct_full),m_pct_full_h,stop_m_pct_full))]  
trajets[,stop_m_pct_full_av:=(ifelse(is.na(stop_m_pct_full_av),m_pct_full_av_h,stop_m_pct_full_av))]  

trajets[,c("id","mois","jour"):=NULL]
trajets[,c("m_total_docks_h","m_docks_in_service_h","m_available_docks_h","m_available_bikes_h","m_pct_full_h","m_pct_full_av_h"):=NULL]

rm(stations_h)
file.remove("D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/fichier_anissa/stations_h.rda")

```


#### Aggrégation par trajet

```{r}
  # Identifiant trajet unique
trajets[,  trajet_id:=paste0("From ",from_station_id, " to ", to_station_id)]

trajets[,c("from_station_name","to_station_name",
           "start_mois","start_jour","start_hr","stop_hr","stop_jour","stop_mois"):=NULL]
```



```{r}
setorderv(trajets,c("trajet_id","wday_n"))

trajets <- trajets[,nb_trips:=.N,by=.(trajet_id,wday_n)]
trajets <- trajets[,pct_abonne:=round(mean(abonne),3)*100,by=.(trajet_id,wday_n)]

trajets <- trajets[,moy_distHav:=round(mean(distHav)),by=.(trajet_id,wday_n)]
trajets <- trajets[,moy_trip_duration:=round(mean(trip_duration)/60),by=.(trajet_id,wday_n)]

trajets <- trajets[,start_moy_total_docks:=round(mean(start_m_total_docks)),by=.(trajet_id,wday_n)]
trajets <- trajets[,start_moy_docks_in_service:=round(mean(start_m_docks_in_service)),by=.(trajet_id,wday_n)]
trajets <- trajets[,start_moy_available_docks:=round(mean(start_m_available_docks)),by=.(trajet_id,wday_n)]
trajets <- trajets[,start_moy_available_bikes:=round(mean(start_m_available_bikes)),by=.(trajet_id,wday_n)]
trajets <- trajets[,start_moy_pct_full:=round(mean(start_m_pct_full)),by=.(trajet_id,wday_n)]
trajets <- trajets[,start_moy_pct_full_av:=round(mean(start_m_pct_full_av)),by=.(trajet_id,wday_n)]

trajets <- trajets[,stop_moy_total_docks:=round(mean(stop_m_total_docks)),by=.(trajet_id,wday_n)]
trajets <- trajets[,stop_moy_docks_in_service:=round(mean(stop_m_docks_in_service)),by=.(trajet_id,wday_n)]
trajets <- trajets[,stop_moy_available_docks:=round(mean(stop_m_available_docks)),by=.(trajet_id,wday_n)]
trajets <- trajets[,stop_moy_available_bikes:=round(mean(stop_m_available_bikes)),by=.(trajet_id,wday_n)]
trajets <- trajets[,stop_moy_pct_full:=round(mean(stop_m_pct_full)),by=.(trajet_id,wday_n)]
trajets <- trajets[,stop_moy_pct_full_av:=round(mean(stop_m_pct_full_av)),by=.(trajet_id,wday_n)]

```


Suppression des variables inutiles

```{r}
trajets[,c("abonne","trip_duration","distHav","start_m_total_docks","start_m_docks_in_service","start_m_available_docks","start_m_available_bikes","start_m_pct_full","start_m_pct_full_av","stop_m_total_docks","stop_m_docks_in_service","stop_m_available_docks","stop_m_available_bikes","stop_m_pct_full","stop_m_pct_full_av"):=NULL]

trajets<-unique(trajets)
setorderv(trajets,c("trajet_id","wday_n"))
```



```{r}
des_trajets <- unique(trajets[,c("from_station_id","to_station_id","trajet_id")])
des_trajets <- merge(des_trajets,lstations,by.x=c("from_station_id"),by.y="id",all.x=TRUE)
setnames(des_trajets,"latitude","from_latitude")
setnames(des_trajets,"longitude","from_longitude")
setnames(des_trajets,"station_name","from_station_name")
des_trajets <- merge(des_trajets,lstations,by.x=c("to_station_id"),by.y="id",all.x=TRUE)
setnames(des_trajets,"latitude","to_latitude")
setnames(des_trajets,"longitude","to_longitude")
setnames(des_trajets,"station_name","to_station_name")

des_trajets[,c("to_station_id","from_station_id"):=NULL]
rm(lstations)

```


```{r}
trajets[,c("to_station_id","from_station_id"):=NULL]

trajets <- na.omit(trajets) 
    # quelques cas où distance non calculable, valeurs manquantes sur tx de remplissage: on les supprime
    # 390 231 obs à 390 145 trajets

# Tri de la table
setorderv(trajets,c("wday_n","trajet_id"))
```


```{r}
saveRDS(trajets,"D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/fichier_anissa/trajets.rda",compress=FALSE)

saveRDS(des_trajets,"D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/fichier_anissa/des_trajets.rda",compress=FALSE)

```
