==> 16_Customers_trips_clustering_final.Rmd <==
<<<<<<< HEAD
---
title: "1_Stat_des_customers_2017"
output: html_document
---
Faire une variante ? 1 seul trajet / > 1 trajet / tous 
Checker si c'est du bruit ou pas.



```{r}
library(data.table)
library(magrittr)
library(tidyverse)  # data manipulation
library(cluster)    # clustering algorithms
library(factoextra) # clustering algorithms & visualization
library(gridExtra)
library(LICORS)
library(leaflet)
library(FactoMineR)
library(dbscan)

```


#### Pr?sentation du programme
# Etape 1 : on exclus les (trajets-jours) avec un seul trajet r?alis?
# Etape 2 : Clustering k-means
# Etape 3 : CAH
# Etape 4 : ACP + CAH


#### Appel de la base

```{r}

  # Appel de la 1?re base de travail : 1 ligne = 1 trajet
  # Pour chaque trajet on a des informations caract?risant ce trajet, les stations d'arriv?e et de d?part
#trajets <- readRDS("D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/fichier_anissa/trajets.rda")
trajets <- readRDS("20_ShinyAppEnd/ChicagoBikes/trajets.rda")

  # Appel de la 2?me base de travail : 1 ligne = 1 station
  # D?crit pour chaque station son nom et sa localisation g?ographique
#des_trajets <- readRDS("D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/fichier_anissa/des_trajets.rda")
des_trajets <- readRDS("20_ShinyAppEnd/ChicagoBikes/des_trajets.rda")
```


Construction de la base de travail

=> Elle d?pendra de 2 param?tres choisis par l'utilisateur : le jour et le type de trajets (unique/pas unique/tous) 

```{r}

# SELECTIONNER LE JOUR (wday_n) : 1: dimanche, 2: lundi... 7: samedi

# SELECTIONNER D'EXCLURE OU NON LES TRAJETS AVEC 1 TRIP (filtre sur la variable nb_trips)
# Tous les trajets    : nb_trips > 0
# Sans trajets unique : nb_trips > 1
# Trajets uniques     : nb_trips = 1

# SELECTIONNER LES VARIABLES (on cr?era le vecteur maselec)
# La s?lection maximale :
#"nb_trips", "pct_abonne","moy_from_bikes","moy_to_bikes","moy_from_trips","moy_to_trips","from_nbst_300m","to_nbst_300m","from_nbst_1km","to_nbst_1km","from_nbst_2km","to_nbst_2km","moy_distHav","moy_trip_duration","start_moy_total_docks","start_moy_docks_in_service","start_moy_available_docks","start_moy_available_bikes","start_moy_pct_full","start_moy_pct_full_av","stop_moy_total_docks","stop_moy_docks_in_service","stop_moy_available_docks","stop_moy_available_bikes","stop_moy_pct_full","stop_moy_pct_full_av"

# Construction des 2 tables de travail
# La premi?re sur laquelle on r?alise l'analyse
# La deuxi?me qui nous permettra d'identifier les trajets 

# Ici exemple : 
# on travaille sur dimanche : jour=1
jour <- 1
# sur tous les trajets (unique,nonunique,tous) : 
unique <- "tous"
# et une sous-selection de variables : maselec

maselec <- c("nb_trips", "pct_abonne",
"moy_from_bikes","moy_to_bikes","moy_from_trips","moy_to_trips",
"from_nbst_300m","to_nbst_300m","from_nbst_1km","to_nbst_1km",
"moy_distHav","moy_trip_duration",
"start_moy_total_docks","start_moy_docks_in_service","start_moy_available_docks","start_moy_pct_full",
"stop_moy_total_docks","stop_moy_docks_in_service","stop_moy_available_docks","stop_moy_pct_full")

if (unique == "tous"){
  trajets_cl<-copy(trajets[wday_n == 1 & nb_trips > 0 ,maselec, with=FALSE])
  trajets_id<-copy(trajets[wday_n == 1 & nb_trips > 0 ,c("trajet_id")])
} else {
    if (unique == "nonunique"){
      trajets_cl<-copy(trajets[wday_n == 1 & nb_trips > 1 ,maselec, with=FALSE])
      trajets_id<-copy(trajets[wday_n == 1 & nb_trips > 1 ,c("trajet_id")])
    }else{
            # Si l'utilisateur a choisit de ne travailler que sur le strajets uniques, 
            # il faut s'assurer que la table ne tienne pas compte de la variable nb_trips
          trajets_cl<-copy(trajets[wday_n == 1 & nb_trips == 1 ,maselec, with=FALSE])
          if (is.na(match("nb_trips",names(trajets_cl)))==FALSE) {
              trajets_cl<-trajets_cl[,-c("nb_trips")]
          }
          trajets_id<-copy(trajets[wday_n==1 & nb_trips == 1 ,c("trajet_id")])
    }
}

```

Question : est-ce qu'uk ets possible de fair evarier les m?thodes sans avoir ? syst?matiquement choisr les varriables ? chaque fois
Et faire en sorte que les deux tables trajets_cl et trajet_id soient modifi?es d?s lors que l'utilisateur modifie le choix de varoables ?


    # Proposer ? l'utilisateur plusieurs choix d'algorithmes
    #     K-means avec l'algorithme "Hartigan-Wong" 
    #     K-means ++ avec s?lection des centro?des de mani?re al?atoires
    #     DBSCAN
    #     ACP + CAH
    
    # Dessiner les classes sur la carte de Chicago plut?t que le cluster


METHODE 1 : K-means

```{r}

# FAIRE CHOISIR LE NOMBRE DE CLASSES : entre 1 et 15 (on cr?era la variable nbcl)
# FAIRE CHOISIR L'ALGORITHME         : "Hartigan-Wong", "Lloyd", "Forgy","MacQueen" (on cr?er la variable algo)

# Exemple avec 6 classes
nbcl <- 6

# Exemple avec l'algo de MacQueen
algo <- "MacQueen"

set.seed(12345)

  # nb de classes comme param?tre de centers
km_hw<- kmeans(scale(trajets_cl,center=T,scale=T),centers=nbcl,iter.max=2000, algorithm = algo,nstart=50)

# Ajout des coordonn?ees g?ographiques pour la visualisation graphique des classes
trajets_res <- cbind.data.frame(trajets_cl,trajets_id,cl_km_hw=factor(km_hw$cluster))
trajets_res <- merge(trajets_res,des_trajets,by=c("trajet_id"),all.x=TRUE)
rm(km_hw)



# Mise ne forme des graphiques : j'aimerais bien que quand on passe la souris sur un point on ait un affichage de la station, du nom du trajet, longutude te latitude

# Taille des classes
print(paste("Effectifs des ",nbcl," classes"))
Eff_trajets <- as.data.table(table(trajets_res[,cl_km_hw]))
Eff_trajets[,Pct:=round((N/sum(N))*100)]
setnames(Eff_trajets,"V1","N? de classe")
setnames(Eff_trajets,"N","Nombre de trajets")
setnames(Eff_trajets,"Pct","%")
print(Eff_trajets)
rm(Eff_trajets)

print(head(trajets_res))
#print(cl_km_hw)


head(trajets_res)
# Je n'ai as r?ussi ? ajouter dans la fonction aes, apr?s colr = cl_km_hw, les infos pour afficher pour chaque point le num?ro de la classe, nom de la station, longitude, latitude
# Ligne de code qui marche pas
# text = paste("Classe: ", cl_km_hw, '<br>',"Station:", to_station_name,'<br>', "Longitude:",to_longitude,'<br>',"Latitude:",to_latitude)


from <- trajets_res[] %>% 
        ggplot() +
        geom_point(data = trajets_res[, .(from_longitude, from_latitude)], aes(from_longitude, from_latitude),
                    color = 'gray') +
        geom_point(aes(from_longitude, from_latitude, color = cl_km_hw, text = paste("Classe: ", cl_km_hw, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
        labs(x="Longitude",y="Latitude",cl_km_hw="Classes")+
        labs(title = paste("Visualisation des ",nbcl,"classes de trajets"), subtitle = "Selon la station de d?part")+
        facet_wrap(~cl_km_hw)

from

to <- trajets_res[] %>% 
      ggplot() +
      geom_point(data = trajets_res[, .(to_longitude, to_latitude)], aes(to_longitude, to_latitude),
                 color = 'gray') +
      geom_point(aes(to_longitude, to_latitude, color = cl_km_hw,text = paste("Classe: ", cl_km_hw, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
      labs(x="Longitude",y="Latitude",color="Classes") +
      labs(title = paste("Visualisation des ",nbcl,"classes de trajets"), subtitle = "Selon la station d'arriv?e")+
      facet_wrap(~cl_km_hw)
to   

rm(from,to)
rm(trajets_res)
```



METHODE 2 : K-means ++


```{r}
# FAIRE CHOISIR LE NOMBRE DE CLASSES : entre 1 et 15 (on cr?era la variable nbcl)
# FAIRE CHOISIR L'ALGORITHME         : "Hartigan-Wong", "Lloyd", "Forgy","MacQueen" (on cr?er la variable algo)
# PAR DEFAUT : le choix des centres de d?part est al?atoire

# Exemple avec 6 classes
nbcl <- 6

# Exemple avec l'algo de MacQueen
algo <- "MacQueen"


set.seed(12345)

kmpp_hw <- kmeanspp(scale(trajets_cl,center=T,scale=T), k = nbcl, start = "random", iter.max = 1000, nstart = 20, algorithm=algo)

# Ajout des coordonn?ees g?ographiques pour la visualisation graphique des classes
trajets_res <- cbind.data.frame(trajets_cl,trajets_id,cl_kmpp_hw=factor(kmpp_hw$cluster))
trajets_res <- merge(trajets_res,des_trajets,by=c("trajet_id"),all.x=TRUE)
rm(kmpp_hw)

# Taille des classes
print(paste("Effectifs des ",nbcl," classes"))
Eff_trajets <- as.data.table(table(trajets_res[,cl_kmpp_hw]))
Eff_trajets[,Pct:=round((N/sum(N))*100)]
setnames(Eff_trajets,"V1","N? de classe")
setnames(Eff_trajets,"N","Nombre de trajets")
setnames(Eff_trajets,"Pct","%")
print(Eff_trajets)
rm(Eff_trajets)


# Je n'ai as r?ussi ? ajouter dans la fonction aes, apr?s colr = cl_kmpp_hw, les infos pour afficher pour chaque point le num?ro de la classe, nom de la station, longitude, latitude
# Ligne de code qui marche pas
# text = paste("Classe: ", cl_kmpp_hw, '<br>',"Station:", to_station_name,'<br>', "Longitude:",to_longitude,'<br>',"Latitude:",to_latitude)


from <- trajets_res[] %>% 
        ggplot() +
        geom_point(data = trajets_res[, .(from_longitude, from_latitude)], aes(from_longitude, from_latitude),
                    color = 'gray') +
        geom_point(aes(from_longitude, from_latitude, color = cl_kmpp_hw, text = paste("Classe: ", cl_kmpp_hw, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
        labs(x="Longitude",y="Latitude",cl_kmpp_hw="Classes")+
        labs(title = paste("Visualisation des ",nbcl,"classes de trajets"), subtitle = "Selon la station de d?part")+
        facet_wrap(~cl_kmpp_hw)

from

to <- trajets_res[] %>% 
      ggplot() +
      geom_point(data = trajets_res[, .(to_longitude, to_latitude)], aes(to_longitude, to_latitude),
                 color = 'gray') +
      geom_point(aes(to_longitude, to_latitude, color = cl_kmpp_hw,text = paste("Classe: ", cl_kmpp_hw, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
      labs(x="Longitude",y="Latitude",color="Classes") +
      labs(title = paste("Visualisation des ",nbcl,"classes de trajets"), subtitle = "Selon la station d'arriv?e")+
      facet_wrap(~cl_kmpp_hw)
to   

rm(from,to)
rm(trajets_res)
```




METHODE 3 : DBSCAN

```{r}
# FAIRE CHOISIR LA VALEUR DE minPts (taille min du voisinage d'un point) : ch_minPts
# choix de eps optimum (k plus proche voisin, on prend la valeur de MinPts pour k)
# Pour MintPts : on prend la valeur de 4
# on cherche quel est le epsilon optimum par la m?thode du coude

# Exemple : on prend 7
ch_minPts <- 7

dbscan::kNNdistplot(scale(trajets_cl,center=T,scale=T), k=ch_minPts)
averageDist <- colMeans(dbscan::kNNdist(scale(trajets_cl,center=T,scale=T),k=ch_minPts))
eps_opt<-mean(averageDist)

res_dbscan <- dbscan::dbscan(scale(trajets_cl, center=T, scale=T), eps=eps_opt, minPts=ch_minPts)
#db6 <-fviz_cluster(res_dbscan, geom = "point", data = trajets_cl)

rm(eps_opt,averageDist)

# Ajout des coordonn?ees g?ographiques pour la visualisation graphique des classes
trajets_res <- cbind.data.frame(trajets_cl,trajets_id,cl_dbscan=factor(res_dbscan$cluster))
trajets_res <- merge(trajets_res,des_trajets,by=c("trajet_id"),all.x=TRUE)

# Taille des classes
print(paste("Effectifs des ",max(res_dbscan$cluster)," classes"))
Eff_trajets <- as.data.table(table(trajets_res[,cl_dbscan]))
Eff_trajets[,Pct:=round((N/sum(N))*100)]
setnames(Eff_trajets,"V1","N? de classe")
setnames(Eff_trajets,"N","Nombre de trajets")
setnames(Eff_trajets,"Pct","%")
print(Eff_trajets)
rm(Eff_trajets)


# Je n'ai pas r?ussi ? ajouter dans la fonction aes, apr?s colr = cl_kmpp_hw, les infos pour afficher pour chaque point le num?ro de la classe, nom de la station, longitude, latitude
# Ligne de code qui marche pas
# text = paste("Classe: ", cl_kmpp_hw, '<br>',"Station:", to_station_name,'<br>', "Longitude:",to_longitude,'<br>',"Latitude:",to_latitude)


from <- trajets_res[] %>% 
        ggplot() +
        geom_point(data = trajets_res[, .(from_longitude, from_latitude)], aes(from_longitude, from_latitude),
                    color = 'gray') +
        geom_point(aes(from_longitude, from_latitude, color = cl_dbscan, text = paste("Classe: ", cl_dbscan, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
        labs(x="Longitude",y="Latitude",cl_dbscan="Classes")+
        labs(title = paste("Visualisation des ",max(res_dbscan$cluster),"classes de trajets"), subtitle = "Selon la station de d?part")+
        facet_wrap(~cl_dbscan)

from

to <- trajets_res[] %>% 
      ggplot() +
      geom_point(data = trajets_res[, .(to_longitude, to_latitude)], aes(to_longitude, to_latitude),
                 color = 'gray') +
      geom_point(aes(to_longitude, to_latitude, color = cl_dbscan,text = paste("Classe: ", cl_dbscan, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
      labs(x="Longitude",y="Latitude",color="Classes") +
      labs(title = paste("Visualisation des ",max(res_dbscan$cluster),"classes de trajets"), subtitle = "Selon la station d'arriv?e")+
      facet_wrap(~cl_dbscan)
to   

rm(from,to)
rm(res_dbscan)
rm(trajets_res)
#revoir db
```

METHODE 4 : A venir...
=======
---
title: "1_Stat_des_customers_2017"
output: html_document
---
Faire une variante ? 1 seul trajet / > 1 trajet / tous 
Checker si c'est du bruit ou pas.



```{r}
library(data.table)
library(magrittr)
library(tidyverse)  # data manipulation
library(cluster)    # clustering algorithms
library(factoextra) # clustering algorithms & visualization
library(gridExtra)
library(LICORS)
library(leaflet)
library(FactoMineR)
library(dbscan)
```


#### Présentation du programme
# Etape 1 : on exclus les (trajets-jours) avec un seul trajet réalisé
# Etape 2 : Clustering k-means
# Etape 3 : CAH
# Etape 4 : ACP + CAH


#### Appel de la base

```{r}

  # Appel de la 1ère base de travail : 1 ligne = 1 trajet
  # Pour chaque trajet on a des informations caractérisant ce trajet, les stations d'arrivée et de départ
trajets <- readRDS("D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/fichier_anissa/trajets.rda")

  # Appel de la 2ème base de travail : 1 ligne = 1 station
  # Décrit pour chaque station son nom et sa localisation géographique
des_trajets <- readRDS("D:/Formations/CEPE_2018_Formation_DataScience/Projet_formation/fichier_anissa/des_trajets.rda")

```


Construction de la base de travail

=> Elle dépendra de 2 paramètres choisis par l'utilisateur : le jour et le type de trajets (unique/pas unique/tous) 

```{r}

# SELECTIONNER LE JOUR (wday_n) : 1: dimanche, 2: lundi... 7: samedi

# SELECTIONNER D'EXCLURE OU NON LES TRAJETS AVEC 1 TRIP (filtre sur la variable nb_trips)
# Tous les trajets    : nb_trips > 0
# Sans trajets unique : nb_trips > 1
# Trajets uniques     : nb_trips = 1

# SELECTIONNER LES VARIABLES (on créera le vecteur maselec)
# La sélection maximale :
#"nb_trips", "pct_abonne","moy_from_bikes","moy_to_bikes","moy_from_trips","moy_to_trips","from_nbst_300m","to_nbst_300m","from_nbst_1km","to_nbst_1km","from_nbst_2km","to_nbst_2km","moy_distHav","moy_trip_duration","start_moy_total_docks","start_moy_docks_in_service","start_moy_available_docks","start_moy_available_bikes","start_moy_pct_full","start_moy_pct_full_av","stop_moy_total_docks","stop_moy_docks_in_service","stop_moy_available_docks","stop_moy_available_bikes","stop_moy_pct_full","stop_moy_pct_full_av"

# Construction des 2 tables de travail
# La première sur laquelle on réalise l'analyse
# La deuxième qui nous permettra d'identifier les trajets 

# Ici exemple : 
# on travaille sur dimanche : jour=1
jour <- 1
# sur tous les trajets (unique,nonunique,tous) : 
unique <- "tous"
# et une sous-selection de variables : maselec

maselec <- c("nb_trips", "pct_abonne",
"moy_from_bikes","moy_to_bikes","moy_from_trips","moy_to_trips",
"from_nbst_300m","to_nbst_300m","from_nbst_1km","to_nbst_1km",
"moy_distHav","moy_trip_duration",
"start_moy_total_docks","start_moy_docks_in_service","start_moy_available_docks","start_moy_pct_full",
"stop_moy_total_docks","stop_moy_docks_in_service","stop_moy_available_docks","stop_moy_pct_full")

if (unique == "tous"){
  trajets_cl<-copy(trajets[wday_n == 1 & nb_trips > 0 ,maselec, with=FALSE])
  trajets_id<-copy(trajets[wday_n == 1 & nb_trips > 0 ,c("trajet_id")])
} else {
    if (unique == "nonunique"){
      trajets_cl<-copy(trajets[wday_n == 1 & nb_trips > 1 ,maselec, with=FALSE])
      trajets_id<-copy(trajets[wday_n == 1 & nb_trips > 1 ,c("trajet_id")])
    }else{
            # Si l'utilisateur a choisit de ne travailler que sur le strajets uniques, 
            # il faut s'assurer que la table ne tienne pas compte de la variable nb_trips
          trajets_cl<-copy(trajets[wday_n == 1 & nb_trips == 1 ,maselec, with=FALSE])
          if (is.na(match("nb_trips",names(trajets_cl)))==FALSE) {
              trajets_cl<-trajets_cl[,-c("nb_trips")]
          }
          trajets_id<-copy(trajets[wday_n==1 & nb_trips == 1 ,c("trajet_id")])
    }
}

```

Question : est-ce qu'uk ets possible de fair evarier les méthodes sans avoir à systématiquement choisr les varriables à chaque fois
Et faire en sorte que les deux tables trajets_cl et trajet_id soient modifiées dès lors que l'utilisateur modifie le choix de varoables ?


    # Proposer à l'utilisateur plusieurs choix d'algorithmes
    #     K-means avec l'algorithme "Hartigan-Wong" 
    #     K-means ++ avec sélection des centroïdes de manière aléatoires
    #     DBSCAN
    #     ACP + CAH
    
    # Dessiner les classes sur la carte de Chicago plutôt que le cluster


METHODE 1 : K-means

```{r}

# FAIRE CHOISIR LE NOMBRE DE CLASSES : entre 1 et 15 (on créera la variable nbcl)
# FAIRE CHOISIR L'ALGORITHME         : "Hartigan-Wong", "Lloyd", "Forgy","MacQueen" (on créer la variable algo)

# Exemple avec 6 classes
nbcl <- 6

# Exemple avec l'algo de MacQueen
algo <- "MacQueen"

set.seed(12345)

  # nb de classes comme paramètre de centers
km_hw<- kmeans(scale(trajets_cl,center=T,scale=T),centers=nbcl,iter.max=2000, algorithm = algo,nstart=50)

# Ajout des coordonnéees géographiques pour la visualisation graphique des classes
trajets_res <- cbind.data.frame(trajets_cl,trajets_id,cl_km_hw=factor(km_hw$cluster))
trajets_res <- merge(trajets_res,des_trajets,by=c("trajet_id"),all.x=TRUE)
rm(km_hw)

# Mise ne forme des graphiques : j'aimerais bien que quand on passe la souris sur un point on ait un affichage de la station, du nom du trajet, longutude te latitude

# Taille des classes
print(paste("Effectifs des ",nbcl," classes"))
Eff_trajets <- as.data.table(table(trajets_res[,cl_km_hw]))
Eff_trajets[,Pct:=round((N/sum(N))*100)]
setnames(Eff_trajets,"V1","N° de classe")
setnames(Eff_trajets,"N","Nombre de trajets")
setnames(Eff_trajets,"Pct","%")
print(Eff_trajets)
rm(Eff_trajets)

# Je n'ai as réussi à ajouter dans la fonction aes, après colr = cl_km_hw, les infos pour afficher pour chaque point le numéro de la classe, nom de la station, longitude, latitude
# Ligne de code qui marche pas
# text = paste("Classe: ", cl_km_hw, '<br>',"Station:", to_station_name,'<br>', "Longitude:",to_longitude,'<br>',"Latitude:",to_latitude)


from <- trajets_res[] %>% 
        ggplot() +
        geom_point(data = trajets_res[, .(from_longitude, from_latitude)], aes(from_longitude, from_latitude),
                    color = 'gray') +
        geom_point(aes(from_longitude, from_latitude, color = cl_km_hw, text = paste("Classe: ", cl_km_hw, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
        labs(x="Longitude",y="Latitude",cl_km_hw="Classes")+
        labs(title = paste("Visualisation des ",nbcl,"classes de trajets"), subtitle = "Selon la station de départ")+
        facet_wrap(~cl_km_hw)

from

to <- trajets_res[] %>% 
      ggplot() +
      geom_point(data = trajets_res[, .(to_longitude, to_latitude)], aes(to_longitude, to_latitude),
                 color = 'gray') +
      geom_point(aes(to_longitude, to_latitude, color = cl_km_hw,text = paste("Classe: ", cl_km_hw, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
      labs(x="Longitude",y="Latitude",color="Classes") +
      labs(title = paste("Visualisation des ",nbcl,"classes de trajets"), subtitle = "Selon la station d'arrivée")+
      facet_wrap(~cl_km_hw)
to   

rm(from,to)
rm(trajets_res)
```



METHODE 2 : K-means ++


```{r}
# FAIRE CHOISIR LE NOMBRE DE CLASSES : entre 1 et 15 (on créera la variable nbcl)
# FAIRE CHOISIR L'ALGORITHME         : "Hartigan-Wong", "Lloyd", "Forgy","MacQueen" (on créer la variable algo)
# PAR DEFAUT : le choix des centres de départ est aléatoire

# Exemple avec 6 classes
nbcl <- 6

# Exemple avec l'algo de MacQueen
algo <- "MacQueen"


set.seed(12345)

kmpp_hw <- kmeanspp(scale(trajets_cl,center=T,scale=T), k = nbcl, start = "random", iter.max = 1000, nstart = 20, algorithm=algo)

# Ajout des coordonnéees géographiques pour la visualisation graphique des classes
trajets_res <- cbind.data.frame(trajets_cl,trajets_id,cl_kmpp_hw=factor(kmpp_hw$cluster))
trajets_res <- merge(trajets_res,des_trajets,by=c("trajet_id"),all.x=TRUE)
rm(kmpp_hw)

# Taille des classes
print(paste("Effectifs des ",nbcl," classes"))
Eff_trajets <- as.data.table(table(trajets_res[,cl_kmpp_hw]))
Eff_trajets[,Pct:=round((N/sum(N))*100)]
setnames(Eff_trajets,"V1","N° de classe")
setnames(Eff_trajets,"N","Nombre de trajets")
setnames(Eff_trajets,"Pct","%")
print(Eff_trajets)
rm(Eff_trajets)


# Je n'ai as réussi à ajouter dans la fonction aes, après colr = cl_kmpp_hw, les infos pour afficher pour chaque point le numéro de la classe, nom de la station, longitude, latitude
# Ligne de code qui marche pas
# text = paste("Classe: ", cl_kmpp_hw, '<br>',"Station:", to_station_name,'<br>', "Longitude:",to_longitude,'<br>',"Latitude:",to_latitude)


from <- trajets_res[] %>% 
        ggplot() +
        geom_point(data = trajets_res[, .(from_longitude, from_latitude)], aes(from_longitude, from_latitude),
                    color = 'gray') +
        geom_point(aes(from_longitude, from_latitude, color = cl_kmpp_hw, text = paste("Classe: ", cl_kmpp_hw, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
        labs(x="Longitude",y="Latitude",cl_kmpp_hw="Classes")+
        labs(title = paste("Visualisation des ",nbcl,"classes de trajets"), subtitle = "Selon la station de départ")+
        facet_wrap(~cl_kmpp_hw)

from

to <- trajets_res[] %>% 
      ggplot() +
      geom_point(data = trajets_res[, .(to_longitude, to_latitude)], aes(to_longitude, to_latitude),
                 color = 'gray') +
      geom_point(aes(to_longitude, to_latitude, color = cl_kmpp_hw,text = paste("Classe: ", cl_kmpp_hw, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
      labs(x="Longitude",y="Latitude",color="Classes") +
      labs(title = paste("Visualisation des ",nbcl,"classes de trajets"), subtitle = "Selon la station d'arrivée")+
      facet_wrap(~cl_kmpp_hw)
to   

rm(from,to)
rm(trajets_res)
```




METHODE 3 : DBSCAN

```{r}
# FAIRE CHOISIR LA VALEUR DE minPts (taille min du voisinage d'un point) : ch_minPts
# choix de eps optimum (k plus proche voisin, on prend la valeur de MinPts pour k)
# Pour MintPts : on prend la valeur de 4
# on cherche quel est le epsilon optimum par la méthode du coude

# Exemple : on prend 7
ch_minPts <- 7

dbscan::kNNdistplot(scale(trajets_cl,center=T,scale=T),k=ch_minPts)
averageDist <- colMeans(dbscan::kNNdist(scale(trajets_cl,center=T,scale=T),k=ch_minPts))
eps_opt<-mean(averageDist)

res_dbscan <- dbscan::dbscan(scale(trajets_cl,center=T,scale=T),eps=eps_opt,minPts=ch_minPts)
#db6 <-fviz_cluster(res_dbscan, geom = "point", data = trajets_cl)

rm(eps_opt,averageDist)

# Ajout des coordonnéees géographiques pour la visualisation graphique des classes
trajets_res <- cbind.data.frame(trajets_cl,trajets_id,cl_dbscan=factor(res_dbscan$cluster))
trajets_res <- merge(trajets_res,des_trajets,by=c("trajet_id"),all.x=TRUE)

# Taille des classes
print(paste("Effectifs des ",max(res_dbscan$cluster)," classes"))
Eff_trajets <- as.data.table(table(trajets_res[,cl_dbscan]))
Eff_trajets[,Pct:=round((N/sum(N))*100)]
setnames(Eff_trajets,"V1","N° de classe")
setnames(Eff_trajets,"N","Nombre de trajets")
setnames(Eff_trajets,"Pct","%")
print(Eff_trajets)
rm(Eff_trajets)


# Je n'ai pas réussi à ajouter dans la fonction aes, après colr = cl_kmpp_hw, les infos pour afficher pour chaque point le numéro de la classe, nom de la station, longitude, latitude
# Ligne de code qui marche pas
# text = paste("Classe: ", cl_kmpp_hw, '<br>',"Station:", to_station_name,'<br>', "Longitude:",to_longitude,'<br>',"Latitude:",to_latitude)


from <- trajets_res[] %>% 
        ggplot() +
        geom_point(data = trajets_res[, .(from_longitude, from_latitude)], aes(from_longitude, from_latitude),
                    color = 'gray') +
        geom_point(aes(from_longitude, from_latitude, color = cl_dbscan, text = paste("Classe: ", cl_dbscan, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
        labs(x="Longitude",y="Latitude",cl_dbscan="Classes")+
        labs(title = paste("Visualisation des ",max(res_dbscan$cluster),"classes de trajets"), subtitle = "Selon la station de départ")+
        facet_wrap(~cl_dbscan)

from

to <- trajets_res[] %>% 
      ggplot() +
      geom_point(data = trajets_res[, .(to_longitude, to_latitude)], aes(to_longitude, to_latitude),
                 color = 'gray') +
      geom_point(aes(to_longitude, to_latitude, color = cl_dbscan,text = paste("Classe: ", cl_dbscan, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
      labs(x="Longitude",y="Latitude",color="Classes") +
      labs(title = paste("Visualisation des ",max(res_dbscan$cluster),"classes de trajets"), subtitle = "Selon la station d'arrivée")+
      facet_wrap(~cl_dbscan)
to   

rm(from,to)
rm(res_dbscan)
rm(trajets_res)
#revoir db
```

METHODE 4 : ALGORITHME CLARA (PAM ne fonctionne pas bien sur grands volumes)

```{r}

# FAIRE CHOISIR LE NOMBRE DE CLASSES : entre 1 et 15 (on créera la variable nbcl)

# FAIRE CHOISIR LA DISTANCE METRIQUE : "euclidean", "manhattan", "jaccard" (on créera la variable distm)

distm <- "euclidean"

# Exemple avec 6 classes
nbcl <- 6


set.seed(12345)

clara <- clara(scale(trajets_cl,center=T,scale=T),metric = distm, pamLike =TRUE,stand=TRUE, samples=500,nbcl)

# Ajout des coordonnéees géographiques pour la visualisation graphique des classes
trajets_res <- cbind.data.frame(trajets_cl,trajets_id,cl_clara=factor(clara$clustering))
trajets_res <- merge(trajets_res,des_trajets,by=c("trajet_id"),all.x=TRUE)
rm(clara)

# Mise ne forme des graphiques : j'aimerais bien que quand on passe la souris sur un point on ait un affichage de la station, du nom du trajet, longutude te latitude

# Taille des classes
print(paste("Effectifs des ",nbcl," classes"))
Eff_trajets <- as.data.table(table(trajets_res[,cl_clara]))
Eff_trajets[,Pct:=round((N/sum(N))*100)]
setnames(Eff_trajets,"V1","N° de classe")
setnames(Eff_trajets,"N","Nombre de trajets")
setnames(Eff_trajets,"Pct","%")
print(Eff_trajets)
rm(Eff_trajets)

# Je n'ai as réussi à ajouter dans la fonction aes, après colr = cl_clara, les infos pour afficher pour chaque point le numéro de la classe, nom de la station, longitude, latitude
# Ligne de code qui marche pas
# text = paste("Classe: ", cl_clara, '<br>',"Station:", to_station_name,'<br>', "Longitude:",to_longitude,'<br>',"Latitude:",to_latitude)

from <- trajets_res[] %>% 
        ggplot() +
        geom_point(data = trajets_res[, .(from_longitude, from_latitude)], aes(from_longitude, from_latitude),
                    color = 'gray') +
        geom_point(aes(from_longitude, from_latitude, color = cl_clara, text = paste("Classe: ", cl_clara, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
        labs(x="Longitude",y="Latitude",color="Classes")+
        labs(title = paste("Visualisation des ",nbcl,"classes de trajets"), subtitle = "Selon la station de départ")+
        facet_wrap(~cl_clara)

from

to <- trajets_res[] %>% 
      ggplot() +
      geom_point(data = trajets_res[, .(to_longitude, to_latitude)], aes(to_longitude, to_latitude),
                 color = 'gray') +
      geom_point(aes(to_longitude, to_latitude, color = cl_clara,text = paste("Classe: ", cl_clara, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
      labs(x="Longitude",y="Latitude",color="Classes") +
      labs(title = paste("Visualisation des ",nbcl,"classes de trajets"), subtitle = "Selon la station d'arrivée")+
      facet_wrap(~cl_clara)
to   

rm(from,to)
rm(trajets_res)
rm(distm,nbcl)
```

METHODE 5 : 

```{r}
# A TESTER par PC de Magda
# --------------------------------------------------------------------------------------

# FAIRE CHOISIR LE NOMBRE DE CLASSES : entre 1 et 15 (on créera la variable nbcl)
# FAIRE CHOISIR LE NOMBRES D'AXES A CONSERVER : 50%, 60%, 70%, 80%, 90%, 100% (on créera la # variable choixvp qui variera entre 1 (pour 50%) et 11 pour 100% 
#  1 pour 50%, 2 pour 55%, 3 pour 60%, 4 pour 65%...

# Exemple avec 6 classes
nbcl <- 6

# Exemple avec choix de garder 90% des axes
choixvp <- 9

set.seed(12345)

# Nombre d'axes issues de l'ACP à conserver pour la CAH
# On réalise l'ACP avec l'ensemble des axes et on récupère la distribution des vp
pca <- PCA(scale(trajets_cl,center=T,scale=T),ncp=Inf,graph=FALSE)

vp <- as.data.table(pca$eig[,3])

vp[,obs:=as.numeric(rownames(vp))]
vp[,lag_Cumvp := shift(.SD, 1, 0, "lag"), .SDcols=c("V1")]
vp[,Cumvp := V1]
vp[,c("V1"):=NULL]
vp[,vp50 := ( Cumvp > 50 & lag_Cumvp  <= 50) ] 
vp[,vp55 := ( Cumvp > 55 & lag_Cumvp  <= 55) ] 
vp[,vp60 := ( Cumvp > 60 & lag_Cumvp  <= 60) ]
vp[,vp65 := ( Cumvp > 65 & lag_Cumvp  <= 65) ] 
vp[,vp70 := ( Cumvp > 70 & lag_Cumvp  <= 70) ]
vp[,vp75 := ( Cumvp > 75 & lag_Cumvp  <= 75) ] 
vp[,vp80 := ( Cumvp > 80 & lag_Cumvp  <= 80) ] 
vp[,vp85 := ( Cumvp > 85 & lag_Cumvp  <= 85) ]
vp[,vp90 := ( Cumvp > 90 & lag_Cumvp  <= 90) ] 
vp[,vp95 := ( Cumvp > 95 & lag_Cumvp  <= 95) ]


nbvp <- switch (choixvp,
    "1" = vp[vp50 == TRUE,obs],
    "2" = vp[vp55 == TRUE,obs],
    "3" = vp[vp60 == TRUE,obs],
    "4" = vp[vp65 == TRUE,obs],
    "5" = vp[vp70 == TRUE,obs],
    "6" = vp[vp75 == TRUE,obs],
    "7" = vp[vp80 == TRUE,obs],
    "8" = vp[vp85 == TRUE,obs],
    "9" = vp[vp90 == TRUE,obs],
    "10" = vp[vp95 == TRUE,obs],
    "11" = vp[vp60 == TRUE,obs]
)



# Mise en oeuvre de la Cah

# On réalise l'ACP avec l'ensemble des axes en ne conservant que le nb d'axes choisis (paramètre nbvp)
pca <- PCA(scale(trajets_cl,center=T,scale=T),ncp=nbvp,graph=FALSE)

# Pour la CAH on donne en paramètre le nombre de classes (nbcl), le nb d'axes à conserver (nbvp)
cah <- HCPC(pca, nb.clust = nbcl,consol = FALSE,graph=FALSE,method="ward",iter.max = 1000)
cah_clopt <- HCPC(pca, min=1,max=15,consol = FALSE,graph=FALSE,method="ward",iter.max = 1000)

print(paste("Inertie totale conservée : ",vp[obs==nbvp,Cumvp]))
print(paste("Nombre de classes optimales suggéré : ",cah_clopt$call$t$nb.clust))
rm(cah_clopt)
rm(choixvp,nbvp,vp)

# Ajout des coordonnéees géographiques pour la visualisation graphique des classes
trajets_res <- cbind.data.frame(trajets_cl,trajets_id,cl_cah=factor(cah$data.clust$clust))
trajets_res <- merge(trajets_res,des_trajets,by=c("trajet_id"),all.x=TRUE)
rm(pca, cah)

# Mise en forme des graphiques : j'aimerais bien que quand on passe la souris sur un point on ait un affichage de la station, du nom du trajet, longutude te latitude

# Taille des classes
print(paste("Effectifs des ",nbcl," classes"))
Eff_trajets <- as.data.table(table(trajets_res[,cl_cah]))
Eff_trajets[,Pct:=round((N/sum(N))*100)]
setnames(Eff_trajets,"V1","N° de classe")
setnames(Eff_trajets,"N","Nombre de trajets")
setnames(Eff_trajets,"Pct","%")
print(Eff_trajets)
rm(Eff_trajets)

# Je n'ai as réussi à ajouter dans la fonction aes, après colr = cl_pca, les infos pour afficher pour chaque point le numéro de la classe, nom de la station, longitude, latitude
# Ligne de code qui marche pas
# text = paste("Classe: ", cl_pca, '<br>',"Station:", to_station_name,'<br>', "Longitude:",to_longitude,'<br>',"Latitude:",to_latitude)


from <- trajets_res[] %>% 
        ggplot() +
        geom_point(data = trajets_res[, .(from_longitude, from_latitude)], aes(from_longitude, from_latitude),
                    color = 'gray') +
        geom_point(aes(from_longitude, from_latitude, color = cl_cah, text = paste("Classe: ", cl_cah, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
        labs(x="Longitude",y="Latitude",color="Classes")+
        labs(title = paste("Visualisation des ",nbcl,"classes de trajets"), subtitle = "Selon la station de départ")+
        facet_wrap(~cl_cah)

from

to <- trajets_res[] %>% 
      ggplot() +
      geom_point(data = trajets_res[, .(to_longitude, to_latitude)], aes(to_longitude, to_latitude),
                 color = 'gray') +
      geom_point(aes(to_longitude, to_latitude, color = cl_cah,text = paste("Classe: ", cl_cah, '<br>',"Station:", from_station_name,'<br>', "Longitude:",from_longitude,'<br>',"Latitude:",from_latitude))) +
      labs(x="Longitude",y="Latitude",color="Classes") +
      labs(title = paste("Visualisation des ",nbcl,"classes de trajets"), subtitle = "Selon la station d'arrivée")+
      facet_wrap(~cl_cah)
to   

rm(from,to)
rm(trajets_res)
rm(cl_cah,nbcl)
```

>>>>>>> 81d8b302080ca42464f4c4728ff2bf0bd0cfe6d0
